### 2017-01-25 #93 @HajkD
biomartr package
The `biomartr` package provides researchers with a useful tool for the efficient, straightforward and reproducible handling of large-scale meta-genomic data from NCBI and ENSEMBL databases and intuitive organism centered interface functions for retrieving functional annotation information from the BioMart database. In general, the package promotes computational reproducibility in genomics studies.

```
Package: biomartr
Title: Genomic Data Retrieval with R
Version: 0.3.0
Author: Hajk-Georg Drost
Maintainer: Hajk-Georg Drost <hgd23@cam.ac.uk>
Description: Perform meta-genomic data retrieval and functional annotation retrieval with
    R.
NeedsCompilation: yes
    biomaRt,
    Biostrings,
    stringi,
    data.table (>= 1.9.4),
    readr (>= 0.2.2),
    downloader (>= 0.3),
    RCurl (>= 1.95-4.5),
    XML (>= 3.98-1.1),
    stringr (>= 0.6.2)
    knitr (>= 1.6),
    rmarkdown (>= 0.3.3),
    devtools (>= 1.6.1),
```

https://github.com/HajkD/biomartr

Life scientists working with genomic data.

The R packages `seqinr` and `biomaRt` were previous attempts to provide parts of the `biomartr` functionality. The `seqinr` package aims to automate sequence retrieval in R but is not designed for meta-genomic approaches and does not include functional annotation. The `biomaRt` package aims to provide functional annotation methods but these are also not designed for meta-genomic approaches and are not easy to use for non-programming experts. The major advantage of `biomartr` is that it does not require profound programming expertise and vastly extends the functionality of both packages: `seqinr` and `biomaRt`. It is optimized to handle large-scale genomic and meta-genomic data using simple and straightforward commands that are useful for life scientists. The functionality provided by `biomartr` aims to promote computational reproducibility among life scientists without profound programming expertise. This is not achieved by the packages `seqinr` and `biomaRt` which are designed for bioinformaticians and other computationally trained researchers.

- [x] Do you intend for this package to go on CRAN?  (is already on CRAN)

Is the first submission.

None declared.
