### 2017-01-28 #93 @sckott
### Editor checks:
- [ ] **Archive** (JOSS only, may be post-review): NA
- [ ] **Version** (JOSS only, may be post-review): NA
---

#### Editor comments

Thanks for your submission!

Currently seeking reviewers.

Reviewers: xxx
Due date: xxx
