### 2017-02-09 5a35083
adapt documentation to new roxygen standard
