+++ man/organismFilters.Rd
@@ -50,9 +50,6 @@ head(organismFilters("Homo sapiens", topic = "id"), 20)
 
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \references{
 \url{http://biomart.org/}
 
@@ -69,4 +66,6 @@ Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21,
 \code{\link{organismBM}}, \code{\link{organismAttributes}}, \code{\link{getAttributes}},
 \code{\link{getDatasets}}, \code{\link{getMarts}}
 }
-
+\author{
+Hajk-Georg Drost
+}
