+++ man/download.database.Rd
@@ -32,10 +32,9 @@ named \code{database} (default) is created and the correspondning database then 
 
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \seealso{
 \code{\link{download.database.all}}, \code{\link{listDatabases}}
 }
-
+\author{
+Hajk-Georg Drost
+}
