+++ man/getSubgroups.Rd
@@ -28,10 +28,9 @@ getKingdoms(db = "genbank")
 getSubgroups(kingdom = "vertebrate_mammalian",db = "genbank")
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \seealso{
 \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}
 }
-
+\author{
+Hajk-Georg Drost
+}
