+++ man/refseqOrganisms.Rd
@@ -13,4 +13,3 @@ proteomes, and CDS files are stored on the NCBI refseq server.
 \author{
 Hajk-Georg Drost
 }
-
