### 2017-02-09 4c12031
update DESCRIPTION for CRAN submission
