### 2017-02-10 f29ea00
update R package version which is now on CRAN
