+++ FAQs.md
@@ -1,0 +1,3 @@
+# FAQs
+
+- [How can I retrieve GO terms for human genes using the ensembl gene IDs?](https://github.com/HajkD/biomartr/issues/5)
