+++ FAQs.md
@@ -1,3 +1,3 @@
 # FAQs
 
-- [How can I retrieve GO terms for human genes using the ensembl gene IDs?](https://github.com/HajkD/biomartr/issues/5)
+- [How can I retrieve GO terms for human genes using ensembl gene IDs?](https://github.com/HajkD/biomartr/issues/5)
