+++ FAQs.md
@@ -1,3 +1,4 @@
 # FAQs
 
 - [How can I retrieve GO terms for human genes using ensembl gene IDs?](https://github.com/HajkD/biomartr/issues/5)
+- [How can I download coding sequences or protein sequences for all sequenced species?](https://www.biostars.org/p/9202/#235962)
