+++ FAQs.md
@@ -2,3 +2,4 @@
 
 - [How can I retrieve GO terms for human genes using ensembl gene IDs?](https://github.com/HajkD/biomartr/issues/5)
 - [How can I download coding sequences or protein sequences for all sequenced species?](https://www.biostars.org/p/9202/#235962)
+- [Where can I Download the human reference genome in fasta format?](https://www.biostars.org/p/1796/#236039)
