### 2017-02-16 #93 @sckott
moving ahead with @naupaka as reviewer - haven't been able to get anyone else
