### 2017-02-19 fdae617
group name retrieval is now based on listGroups()
