### 2017-02-19 a0bbae9
listGroups() is now designed to better subset groups and also now returns organism names belonging to a specific subgroup
