### 2017-02-19 14cb0ce
new helper func
