### 2017-02-19 3417792
update roxygen based documentations
