+++ man/getGFF.Rd
@@ -55,9 +55,10 @@ getGFF( db       = "genbank",
 }
 
 }
-\seealso{
-\code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGenome}}, \code{\link{meta.retrieval}}
-}
 \author{
 Hajk-Georg Drost
 }
+\seealso{
+\code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGenome}}, \code{\link{meta.retrieval}}
+}
+
