+++ man/getMetaGenomeAnnotations.Rd
@@ -27,9 +27,10 @@ listMetaGenomes()
 getMetaGenomeAnnotations(name = "human gut metagenome")
 }
 }
-\seealso{
-\code{\link{getMetaGenomes}}, \code{\link{listMetaGenomes}}, \code{\link{getGFF}}
-}
 \author{
 Hajk-Georg Drost
 }
+\seealso{
+\code{\link{getMetaGenomes}}, \code{\link{listMetaGenomes}}, \code{\link{getGFF}}
+}
+
