+++ man/meta.retrieval.all.Rd
@@ -33,3 +33,4 @@ meta.retrieval.all(db = "genbank", type = "ensemblgenomes")
 \author{
 Hajk-Georg Drost
 }
+
