### 2017-02-19 5cbdfc9
remove function getSubgroups()
