### 2017-02-19 130884b
remove function listSubgroups()
