+++ NAMESPACE
@@ -18,7 +18,6 @@ export(getMetaGenomeAnnotations)
 export(getMetaGenomeSummary)
 export(getMetaGenomes)
 export(getProteome)
-export(getSubgroups)
 export(getSummaryFile)
 export(is.genome.available)
 export(listDatabases)
@@ -26,7 +25,6 @@ export(listGenomes)
 export(listGroups)
 export(listKingdoms)
 export(listMetaGenomes)
-export(listSubgroups)
 export(meta.retrieval)
 export(meta.retrieval.all)
 export(organismAttributes)
