### 2017-02-19 fc20e1a
add new functionality that allows users to retrieve genomes from subgroups of kingdoms
