+++ man/meta.retrieval.Rd
@@ -12,6 +12,8 @@ meta.retrieval(kingdom, group = NULL, db = "refseq", type = "genome",
 e.g. "archaea","bacteria", "fungi", "invertebrate", "plant", "protozoa", "vertebrate_mammalian", or "vertebrate_other".
 Available kingdoms can be retrieved with \code{\link{getKingdoms}}.}
 
+\item{group}{only species belonging to this subgroup will be downloaded. Groups can be retrieved with \code{\link{getGroups}}.}
+
 \item{db}{a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "emsembl"} or \code{db = "ensemblgenomes"}.}
 
 \item{type}{type of sequences that shall be retrieved. Either \code{genome}, \code{proteome}, or \code{CDS}.}
