### 2017-02-19 58e4dcb
update NEWS
