### 2017-02-19 ab1ce77
update NEWS
