+++ R/getGroups.R
@@ -16,7 +16,7 @@
 #' # retrieve subgroups for vertebrate_mammalian available from genbank
 #' getGroups(kingdom = "vertebrate_mammalian",db = "genbank")
 #' }
-#' @seealso \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getKingdoms}}, \code{\link{getSubgroups}}
+#' @seealso \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getKingdoms}}
 #' @export
 
 getGroups <- function(kingdom, db = "refseq") {
