+++ R/getKingdoms.R
@@ -11,7 +11,7 @@
 #' # retrieve kingdoms available from genbank
 #' getKingdoms(db = "genbank")
 #' }
-#' @seealso \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getSubgroups}}, \code{\link{getGroups}}
+#' @seealso \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGroups}}
 #' @export
 
 getKingdoms <- function(db = "refseq"){
