+++ man/getGroups.Rd
@@ -32,6 +32,6 @@ getGroups(kingdom = "vertebrate_mammalian",db = "genbank")
 Hajk-Georg Drost
 }
 \seealso{
-\code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getKingdoms}}, \code{\link{getSubgroups}}
+\code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getKingdoms}}
 }
 
