+++ man/getKingdoms.Rd
@@ -26,6 +26,6 @@ getKingdoms(db = "genbank")
 Hajk-Georg Drost
 }
 \seealso{
-\code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getSubgroups}}, \code{\link{getGroups}}
+\code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGroups}}
 }
 
