### 2017-02-19 f45713b
add examples for new meta.retrieval() functionality
