### 2017-02-19 fbbbd8c
extend examples
