+++ NAMESPACE
@@ -6,6 +6,7 @@ export(download.database.all)
 export(getAttributes)
 export(getCDS)
 export(getDatasets)
+export(getENSEMBLInfo)
 export(getFilters)
 export(getGFF)
 export(getGO)
