### 2017-02-20 14cd290
add new func to perform ENSEMBL Info retrieval
