### 2017-02-20 95b5428
add new func to retrieve ENSEMBLGENOMES info
