+++ NAMESPACE
@@ -3,12 +3,14 @@
 export(biomart)
 export(download.database)
 export(download.database.all)
+export(getAssemblyStats)
 export(getAttributes)
 export(getCDS)
 export(getDatasets)
 export(getENSEMBLGENOMESInfo)
 export(getENSEMBLInfo)
 export(getFilters)
+export(getGENOMEREPORT)
 export(getGFF)
 export(getGO)
 export(getGenome)
@@ -32,6 +34,7 @@ export(meta.retrieval.all)
 export(organismAttributes)
 export(organismBM)
 export(organismFilters)
+export(read_assemblystats)
 export(read_cds)
 export(read_genome)
 export(read_gff)
