### 2017-02-24 b04a527
export internal getGENOMEREPORT() function
