### 2017-02-24 832601c
add new function read_assemblystats()
