+++ R/read_assemblystats.R
@@ -1,0 +1,81 @@
+#' @title Import Genome Assembly Stats File
+#' @description This function reads an organism specific Genome Assembly Stats file that was retrieved with \code{\link{getAssemblyStats}}.
+#' @param file a character string specifying the path to the file storing the Genome Assembly Stats file.
+#' @param type either \code{type = "raw"} to import the entire genome assembly stats file or \code{type = "stats"} to import overall statsitics including all chromosomes, mitochondria and plastids.
+#' @author Hajk-Georg Drost
+#' @details This function takes a string specifying the path to the Genome Assembly Stats file
+#' of interest (e.g. the path returned by \code{\link{getAssemblyStats}}) and imports it.
+#' @seealso \code{\link{getAssemblyStats}}, \code{\link{read_genome}}, \code{\link{read_proteome}}, \code{\link{read_cds}}, \code{\link{read_gff}}
+#' @export
+read_assemblystats <- function(file, type = "raw") {
+    
+    if (!is.element(type, c("raw", "stats")))
+        stop("Please choose a type that is supported byt this function, e.g. type = 'raw' or type = 'stats'.", call. = FALSE)
+    
+    assemblystats_file <- readr::read_delim(
+        file,
+        delim = "\t",
+        col_names = c(
+            "unit_name",
+            "molecule_name",
+            "molecule_type",
+            "sequence_type",
+            "statistic",
+            "value"
+        ),
+        col_types = readr::cols(
+            "unit_name" = readr::col_character(),
+            "molecule_name" = readr::col_character(),
+            "molecule_type" = readr::col_character(),
+            "sequence_type" = readr::col_character(),
+            "statistic" = readr::col_character(),
+            "value" = readr::col_integer()
+        ),
+        comment = "#"
+    )
+    
+    if (type == "stats") {
+        # select all features (including chromosomes, mitochondria, and plastids)
+        assemblystats_file.all.features <- dplyr::filter(assemblystats_file, unit_name == "all", molecule_name == "all")
+        
+        total.length <- dplyr::filter(assemblystats_file.all.features, statistic == "total-length")
+        spanned.gaps <- dplyr::filter(assemblystats_file.all.features, statistic == "spanned-gaps")
+        unspanned.gaps <- dplyr::filter(assemblystats_file.all.features, statistic == "unspanned-gaps")
+        region.count <- dplyr::filter(assemblystats_file.all.features, statistic == "region-count")
+        scaffold.count <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-count")
+        scaffold.N50 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-N50")
+        scaffold.L50 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-L50")
+        scaffold.N75 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-N75")
+        scaffold.N90 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-N90")
+        contig.count <- dplyr::filter(assemblystats_file.all.features, statistic == "contig-count")
+        contig.N50 <- dplyr::filter(assemblystats_file.all.features, statistic == "contig-N50")
+        total.gap.length <- dplyr::filter(assemblystats_file.all.features, statistic == "total-gap-length")
+        molecule.count <- dplyr::filter(assemblystats_file.all.features, statistic == "molecule-count")
+        top.level.count <- dplyr::filter(assemblystats_file.all.features, statistic == "top-level-count")
+        
+        assemblystats_file.all.features.short <-
[…]
