### 2017-02-24 ff9b00f
add new function getAssemblyStats()
