### 2017-02-24 681215c
update NEWS
