### 2017-02-24 6680065
add functionality to retrieve assembly stats files for all species belonging to the same kingdom or subgroup
