### 2017-02-24 986260a
update meta.retrieval() call and documentation
