### 2017-02-24 668fdae
add option to combine assembly stats file
