### 2017-03-01 7f2de47
update NEWS
