### 2017-03-01 dcff563
make compatible with Ensembl Plants, Ensembl Metazoa, Ensembl Protist, and Ensembl Fungi
