### 2017-03-01 21b3012
make compatible with Ensembl Plants, Ensembl Metazoa, Ensembl Protist, and Ensembl Fungi
