+++ R/getAttributes.R
@@ -26,18 +26,58 @@
 getAttributes <- function(mart, dataset){
         
         if ((!is.character(mart)) || (!is.character(dataset)))
-                stop("Please use a character string as mart or dataset.")
-        
-    url <-
-        paste0(
-            "http://www.ensembl.org/biomart/martservice?type=attributes&dataset=",
-            dataset,
-            "&requestid=biomart&mart=",
-            mart,
-            "&virtualSchema=default"
-        )
+                stop("Please use a character string as mart or dataset.", call. = FALSE)
     
-    testContent <- httr::content(httr::GET(url), as = "text")
+    if (stringr::str_detect(mart, "ENSEMBL"))
+        # connect to BioMart API
+        url <-
+            paste0(
+                "http://ensembl.org/biomart/martservice?type=attributes&dataset=",
+                dataset,
+                "&requestid=biomart&mart=",
+                mart
+            )    
+    if (stringr::str_detect(mart, "plants"))
+        # connect to BioMart API
+        url <-
+            paste0(
+                "http://plants.ensembl.org/biomart/martservice?type=attributes&dataset=",
+                dataset,
+                "&requestid=biomart&mart=",
+                mart
+            )    
+    if (stringr::str_detect(mart, "fung"))
+        # connect to BioMart API
+        url <-
+            paste0(
+                "http://fungi.ensembl.org/biomart/martservice?type=attributes&dataset=",
+                dataset,
+                "&requestid=biomart&mart=",
+                mart
+            )
+    
+    if (stringr::str_detect(mart, "protist"))
+        # connect to BioMart API
+        url <-
+            paste0(
+                "http://protist.ensembl.org/biomart/martservice?type=attributes&dataset=",
+                dataset,
+                "&requestid=biomart&mart=",
+                mart
+            )    
+    if (stringr::str_detect(mart, "metazoa"))
+        # connect to BioMart API
+        url <-
+            paste0(
+                "http://metazoa.ensembl.org/biomart/martservice?type=attributes&dataset=",
+                dataset,
+                "&requestid=biomart&mart=",
+                mart
+            )    
+        
+    
+    
+    testContent <- httr::content(httr::GET(url), as = "text", encoding = "UTF-8")
     if (testContent == "Attribute 'mains' does not exist\n") {
         warning("No attributes were available for mart = ",mart," and dataset = ",dataset,".", call. = FALSE)
         attrBioMart <- data.frame(name = "NA", description = "NA")
@@ -52,7 +92,7 @@ getAttributes <- function(mart, dataset){
     # extract attribute name and attribute description
     suppressWarnings(rawDF <-
                          do.call("rbind", apply(as.data.frame(strsplit(
-                             httr::content(xmlContentAttributes, as = "text"), "\n"
+                             httr::content(xmlContentAttributes, as = "text", encoding = "UTF-8"), "\n"
                          )), 1, function(x)
                              unlist(strsplit(x, "\t")))))
     
