### 2017-03-01 88f9ce4
make compatible with Ensembl Plants, Ensembl Metazoa, Ensembl Protist, and Ensembl Fungi
