### 2017-03-01 0c46566
make compatible with Ensembl Plants, Ensembl Metazoa, Ensembl Protist, and Ensembl Fungi
