### 2017-03-01 b142ee5
make compatible with Ensembl Plants, Ensembl Metazoa, Ensembl Protist, and Ensembl Fungi
