+++ R/getMarts.R
@@ -13,31 +13,14 @@
 #' @seealso \code{\link{getDatasets}}, \code{\link{getAttributes}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 #' @export
 getMarts <- function(){
+    
+    submarts <- c("ensembl", "plants", "fungi", "protists", "metazoa")
+    submarts.df <- vector("list", length(submarts))
+    for (i in seq_along(submarts)) {
+        submarts.df[i] <- list(getSubMarts(submarts[i]))
+    }
         
-        # connect to BioMart API
-        biomartPage <- httr::handle("http://www.ensembl.org:80/biomart/martservice?type=registry&requestid=biomart")
-        xmlContentMarts <- httr::GET(handle = biomartPage)
-        
-        # test whether or not a connection could be established
-        httr::stop_for_status(xmlContentMarts)
-        
-        # parse Mart information
-        doc <- XML::xmlTreeParse(xmlContentMarts, useInternal = TRUE)
-        rootNode <- XML::xmlRoot(doc)
-        
-        # extract available databases
-        databases <- as.data.frame(XML::xmlSApply(rootNode, function(x) XML::xmlGetAttr(x,"name")))
-        
-        # extract available database versions
-        displayNames <- as.data.frame(XML::xmlSApply(rootNode, function(x) XML::xmlGetAttr(x,"displayName")))
-        
-        # ectract information whether or not the corresponding database is visible 
-        visible <- as.data.frame(XML::xmlSApply(rootNode, function(x) XML::xmlGetAttr(x,"visible")))
-        
-        dbBioMart <- data.frame(mart = databases[ , 1], version = displayNames[ , 1], visible = visible[ , 1])
-        
-        return(dplyr::filter(dbBioMart, visible != "0")[ , c("mart","version")])
-               
+    return(dplyr::bind_rows(submarts.df))
 }
 
 
