### 2017-03-01 b63b59e
new helper function to make organism*() functions compatible with Ensembl Plants, Ensembl Metazoa, Ensembl Protist, and Ensembl Fungi
