+++ R/biomart.R
@@ -34,7 +34,7 @@
 #'         mart       = "plants_mart", 
 #'         dataset    = "athaliana_eg_gene",
 #'         attributes = c("start_position","end_position","description"),
-#'         filters    = "tair_locus") 
+#'         filters    = "ensembl_gene_id") 
 #' 
 #' }
 #' 
