+++ tests/testthat/test-getAttributes.R
@@ -13,7 +13,7 @@ test_that("The getAttributes() interface works properly..",{
         marts <- getMarts()[ , "mart"]
         ensemb_mart <- which(stringr::str_match(marts, "ENSEMBL_MART_ENSEMBL") != "NA")
         
-        attrib <- getAttributes(mart = as.character(marts[ensemb_mart]), dataset = "hsapiens_gene_ensembl")
+        attrib <- getAttributes(mart = as.character(marts$mart[ensemb_mart]), dataset = "hsapiens_gene_ensembl")
         
         expect_true(equal_lists(as.list(attrib[1 , ]), list(name = "ensembl_gene_id",
                                                             description = "Gene ID")))
