+++ tests/testthat/test-getDatasets.R
@@ -13,7 +13,7 @@ test_that("The getDatasets() interface works properly..",{
         marts <- getMarts()[ , "mart"]
         ex_mart <- which(stringr::str_match(marts, "ENSEMBL_MART_ENSEMBL") != "NA")
         
-        attrib <- getDatasets(mart = as.character(marts[ex_mart]))
+        attrib <- getDatasets(mart = as.character(marts$mart[ex_mart]))
         
         expect_true(equal_lists(as.list(attrib[1 , ]), list(dataset = "oanatinus_gene_ensembl",
                                                             description = "Platypus genes (OANA5)",
