### 2017-03-02 92f66e7
update global variables
