### 2017-03-02 9aaebce
update global variables
