### 2017-03-02 044a0e7
update global variables
