### 2017-03-02 cee0a40
update global variables
