### 2017-03-02 7e0e3fd
update annotation vignette
