### 2017-03-02 de41b61
update meta genome vignette
