### 2017-03-02 1481f33
update sequence retrieval vignette
