+++ man/getCDS.Rd
@@ -34,10 +34,9 @@ Ath_CDS <- read_cds(file_path, format = "fasta")
 
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \seealso{
 \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
 }
-
+\author{
+Hajk-Georg Drost
+}
