+++ man/organismBM.Rd
@@ -42,9 +42,6 @@ head(organismBM(update = TRUE), 20)
 
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \references{
 \url{http://biomart.org/}
 
@@ -61,4 +58,6 @@ Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21,
 \code{\link{getMarts}}, \code{\link{getDatasets}}, \code{\link{biomart}},
 \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 }
-
+\author{
+Hajk-Georg Drost
+}
