+++ man/getKingdoms.Rd
@@ -22,10 +22,9 @@ getKingdoms(db = "refseq")
 getKingdoms(db = "genbank")
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \seealso{
 \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGroups}}
 }
-
+\author{
+Hajk-Georg Drost
+}
