+++ man/meta.retrieval.all.Rd
@@ -37,4 +37,3 @@ meta.retrieval.all(db = "genbank", type = "ensemblgenomes")
 \author{
 Hajk-Georg Drost
 }
-
