+++ man/getGroups.Rd
@@ -28,10 +28,9 @@ getKingdoms(db = "genbank")
 getGroups(kingdom = "vertebrate_mammalian",db = "genbank")
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \seealso{
 \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getKingdoms}}
 }
-
+\author{
+Hajk-Georg Drost
+}
