+++ man/listDatabases.Rd
@@ -33,10 +33,9 @@ select an appropriate database for download with \code{\link{download.database}}
 # listDatabases(db = "refseq_protein")
 
 }
-\author{
-Hajk-Georg Drost
-}
 \seealso{
 \code{\link{download.database}}
 }
-
+\author{
+Hajk-Georg Drost
+}
