+++ man/organismAttributes.Rd
@@ -53,9 +53,6 @@ head(organismAttributes("Homo sapiens", topic = "homolog"), 20)
 
 }
 }
-\author{
-Hajk-Georg Drost
-}
 \references{
 \url{http://biomart.org/}
 
@@ -72,4 +69,6 @@ Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21,
 \code{\link{organismFilters}}, \code{\link{organismBM}}, \code{\link{biomart}},
 \code{\link[biomaRt]{listAttributes}}
 }
-
+\author{
+Hajk-Georg Drost
+}
