### 2017-03-13 5ec09c6
add new helper function to test if FTP file of query is available or not
