### 2017-03-13 9647135
add ftp file existence check
