### 2017-03-13 330c4df
add ftp file existence check
