+++ R/getCDS.R
@@ -82,6 +82,11 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             local.org <- stringr::str_replace_all(organism,"-","_")
             local.org <- stringr::str_replace_all(organism,"\\/","_")
             
+            if (!exists.ftp.file(url = paste0(FoundOrganism$ftp_path,"/"), file.path = download_url)) {
+                message("Unfortunately no CDS file could be found for organism '",organism,"'. Thus, the download of this organism has been omitted.")
+                return(FALSE) 
+            }
+            
             if (nrow(FoundOrganism) == 1) {
                 tryCatch({
                     utils::capture.output(downloader::download(
