### 2017-03-13 c93b0a8
add ftp file existence check
