### 2017-03-13 48a288c
add ftp file existence check
