### 2017-03-13 d6435ac
add ftp file existence check
