### 2017-03-13 cd7f97b
add ftp file existence check
