### 2017-03-13 73c9f34
add ftp file existence check
