+++ R/getProteome.R
@@ -98,6 +98,11 @@ getProteome <- function(db = "refseq", organism, path = file.path("_ncbi_downloa
             local.org <- stringr::str_replace_all(organism,"-","_")
             local.org <- stringr::str_replace_all(organism,"\\/","_")
             
+            if (!exists.ftp.file(url = paste0(FoundOrganism$ftp_path,"/"), file.path = download_url)) {
+                message("Unfortunately no proteome file could be found for organism '",organism,"'. Thus, the download of this organism has been omitted.")
+                return(FALSE) 
+            }
+            
             if (nrow(FoundOrganism) == 1) {
                 tryCatch({
                     utils::capture.output(downloader::download(
