### 2017-03-13 fd6c5e7
update NEWS
