### 2017-03-13 1e5e85f
add curl package to dependencies
