### 2017-03-13 4e83000
add system sleep to not viotate the NCBI query convention
