+++ NAMESPACE
@@ -40,4 +40,5 @@ export(read_genome)
 export(read_gff)
 export(read_proteome)
 export(refseqOrganisms)
-import(data.table)
+import(Biostrings)
+import(biomaRt)
