### 2017-03-13 bcf030d
import Biostrings package
