### 2017-03-13 fc8c7e5
import Biostrings package
