### 2017-03-13 02e195b
import Biostrings package
