### 2017-03-13 5a1de3e
make examples more generic
