+++ vignettes/Database_Retrieval.Rmd
@@ -25,23 +25,9 @@ The `download.database()` and `download.database.all()` functions implemented in
 
 Before downloading specific databases from NCBI users might want to list available databases. Using the `listDatabases()` function users can retrieve a list of  available databases stored on NCBI.
 
-```r
-library("biomartr")
-
+```{r, eval=TRUE}
 # retrieve a list of available sequence databases at NCBI
-listDatabases(db = "all")
-```
-
-```
- [1] "16SMicrobial"             "cdd_delta"                "cloud"                    "env_nr"
- [5] "env_nt"                   "est"                      "est_human"                "est_mouse"
- [9] "est_others"               "FASTA"                    "gene_info"                "gss"
-[13] "gss_annot"                "htgs"                     "human_genomic"            "human_genomic_transcript"
-[17] "landmark"                 "mouse_genomic_transcript" "nr"                       "nt"
-[21] "other_genomic"            "pataa"                    "patnt"                    "pdbaa"
-[25] "pdbnt"                    "refseq_genomic"           "refseq_protein"           "refseq_rna"
-[29] "refseqgene"               "Representative_Genomes"   "sts"                      "swissprot"
-[33] "taxdb"                    "tsa_nr"                   "tsa_nt"                   "vector"
+biomartr::listDatabases(db = "all")
 ```
 
 However, in case users already know which database they would like to retrieve
@@ -50,142 +36,62 @@ they can filter for the exact files by specifying the NCBI database name. In the
 
 First, the `listDatabases(db = "nr")` allows to list all files corresponding to the `nr` database. 
 
-```r
+```{r, eval=TRUE}
 # show all NCBI nr files
-listDatabases(db = "nr")
+biomartr::listDatabases(db = "nr")
 ```
 
-```
- [1] "nr.00.tar.gz" "nr.01.tar.gz" "nr.02.tar.gz" "nr.03.tar.gz" "nr.04.tar.gz" "nr.05.tar.gz"
- [7] "nr.16.tar.gz" "nr.06.tar.gz" "nr.15.tar.gz" "nr.30.tar.gz" "nr.07.tar.gz" "nr.08.tar.gz"
-[13] "nr.09.tar.gz" "nr.10.tar.gz" "nr.11.tar.gz" "nr.12.tar.gz" "nr.13.tar.gz" "nr.14.tar.gz"
-[19] "nr.28.tar.gz" "nr.29.tar.gz" "nr.31.tar.gz" "nr.17.tar.gz" "nr.18.tar.gz" "nr.19.tar.gz"
-[25] "nr.20.tar.gz" "nr.21.tar.gz" "nr.22.tar.gz" "nr.23.tar.gz" "nr.32.tar.gz" "nr.24.tar.gz"
-[31] "nr.25.tar.gz" "nr.26.tar.gz" "nr.27.tar.gz" "nr.33.tar.gz" "nr.34.tar.gz" "nr.35.tar.gz"
-[37] "nr.36.tar.gz" "nr.37.tar.gz" "nr.38.tar.gz" "nr.39.tar.gz" "nr.40.tar.gz" "nr.41.tar.gz"
-```
 
 The output illustrates that the `NCBI nr` database has been separated into 41 files.
 
 Further examples are:
 
 
-```r
[…]
