+++ NAMESPACE
@@ -42,3 +42,4 @@ export(read_proteome)
 export(refseqOrganisms)
 import(Biostrings)
 import(biomaRt)
+import(curl)
