+++ R/exists.ftp.file.R
@@ -1,3 +1,4 @@
+#' @import curl
 exists.ftp.file <- function(url, file.path) {
     con <- curl::curl(url = url, "rb", curl::new_handle(dirlistonly = TRUE))
     ftp.content <- suppressMessages(readr::read_delim(con, delim = "\n", col_names = FALSE))
