### 2017-03-13 4566a9e
exchange cat() to message()
