### 2017-03-13 625a76c
exchange cat() to message()
