### 2017-03-13 114c5fd
exchange cat() to message() and remove setwd()
