+++ R/getGFF.R
@@ -76,9 +76,7 @@ getGFF <-
                 )
             
             if (nrow(FoundOrganism) == 0) {
-                cat("\n")
-                cat(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
-                cat("\n")
+                message(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
             } else {
                 if (nrow(FoundOrganism) > 1) {
                     warnings(
@@ -198,12 +196,8 @@ getGFF <-
                         "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage '",json.qry.info,"' currently available?", call. = FALSE
                     ))
                 
-                cwd <- getwd()
-                
-                setwd(path)
-                
                 # generate Genome documentation
-                sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+                sink(file.path(path, paste0("doc_",new.organism,"_db_",db,".txt")))
                 
                 cat(paste0("File Name: ", genome.path))
                 cat("\n")
@@ -224,8 +218,6 @@ getGFF <-
                 cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
                 
                 sink()
-                
-                setwd(cwd)
                 
                 print(
                     paste0(
@@ -273,12 +265,8 @@ getGFF <-
                         "The API 'http://rest.ensemblgenomes.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensemblgenomes.org' currently available?", call. = FALSE
                     ))
                 
-                cwd <- getwd()
-                
-                setwd(path)
-                
                 # generate Genome documentation
-                sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+                sink(file.path(path, paste0("doc_",new.organism,"_db_",db,".txt")))
                 
                 cat(paste0("File Name: ", genome.path))
                 cat("\n")
@@ -299,9 +287,7 @@ getGFF <-
                 cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
                 
                 sink()
-                
-                setwd(cwd)
-                
+            
                 print(
                     paste0(
                         "The *.gff annotation file of '",
