### 2017-03-13 9c2d0c0
exchange cat() to message() and remove setwd()
