+++ R/getGenome.R
@@ -78,9 +78,7 @@ getGenome <-
                 )
             
             if (nrow(FoundOrganism) == 0) {
-                cat("\n")
-                cat(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
-                cat("\n")
+                message(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
             } else {
                 if (nrow(FoundOrganism) > 1) {
                     warnings(
@@ -201,12 +199,8 @@ getGenome <-
                         "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensembl.org' currently available?", call. = FALSE
                     ))
                 
-                cwd <- getwd()
-                
-                setwd(path)
-                
                 # generate Genome documentation
-                sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+                sink(file.path(path, paste0("doc_",new.organism,"_db_",db,".txt")))
                 
                 cat(paste0("File Name: ", genome.path))
                 cat("\n")
@@ -227,8 +221,6 @@ getGenome <-
                 cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
                 
                 sink()
-                
-                setwd(cwd)
                 
                 print(
                     paste0(
@@ -275,13 +267,9 @@ getGenome <-
                     stop(
                         "The API 'http://rest.ensemblgenomes.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensemblgenomes.org' currently available?", call. = FALSE
                     ))
-                
-                cwd <- getwd()
-                
-                setwd(path)
-                
+            
                 # generate Genome documentation
-                sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+                sink(file.path(path,paste0("doc_",new.organism,"_db_",db,".txt")))
                 
                 cat(paste0("File Name: ", genome.path))
                 cat("\n")
@@ -302,8 +290,6 @@ getGenome <-
                 cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
                 
                 sink()
-                
-                setwd(cwd)
                 
                 print(
                     paste0(
