### 2017-03-13 9bc1afc
exchange cat() to message() and remove setwd()
