### 2017-03-13 d5c249f
exchange cat() to message()
