### 2017-03-13 0f36309
exchange cat() to message() and remove setwd()
