### 2017-03-13 c6991ab
update vignette
