+++ vignettes/Sequence_Retrieval.Rmd
@@ -518,7 +518,7 @@ submitter: Genome Reference Consortium
 
 In summary, the `getGenome()` and `read_genome()` functions allow users to retrieve genome assemblies by specifying
 the scientific name of the organism of interest and allow them to import the retrieved genome assembly e.g. as `Biostrings` object.
-Thus, users can then perform the [Biostrings notation](https://web.stanford.edu/class/bios221/labs/biostrings/lab_1_biostrings.html) to work with downloaded genomes and can rely on the log
+Thus, users can then perform the `Biostrings notation` to work with downloaded genomes and can rely on the log
 file generated by `getGenome()` to better document the source and version of genome assemblies used for subsequent studies.
 
 Alternatively, users can perform the pipeline logic of the [magrittr](https://github.com/smbache/magrittr) package:
