+++ vignettes/Sequence_Retrieval.Rmd
@@ -958,7 +958,7 @@ submitter: Genome Reference Consortium
 
 In summary, the `getCDS()` and `read_cds()` functions allow users to retrieve CDS files by specifying
 the scientific name of the organism of interest and allow them to import the retrieved CDS file e.g. as `Biostrings` object.
-Thus, users can then perform the [Biostrings notation](https://web.stanford.edu/class/bios221/labs/biostrings/lab_1_biostrings.html) to work with downloaded CDS and can rely on the log
+Thus, users can then perform the `Biostrings notation` to work with downloaded CDS and can rely on the log
 file generated by `getCDS()` to better document the source and version of CDS used for subsequent studies.
 
 Alternatively, users can perform the pipeline logic of the [magrittr](https://github.com/smbache/magrittr) package:
