+++ NEWS.md
@@ -1,3 +1,7 @@
+biomartr 0.5.0
+===========
+
+
 biomartr 0.4.0
 ===========
 
