### 2017-03-14 afc48c8
update README since Version 0.4.0 is on CRAN nonw
