+++ NAMESPACE
@@ -22,6 +22,7 @@ export(getMetaGenomeAnnotations)
 export(getMetaGenomeSummary)
 export(getMetaGenomes)
 export(getProteome)
+export(getRNA)
 export(getSummaryFile)
 export(is.genome.available)
 export(listDatabases)
@@ -39,6 +40,7 @@ export(read_cds)
 export(read_genome)
 export(read_gff)
 export(read_proteome)
+export(read_rna)
 export(refseqOrganisms)
 import(Biostrings)
 import(biomaRt)
