### 2017-03-15 ee5b1c4
update NEWS
