+++ R/getCDS.R
@@ -19,7 +19,7 @@
 #' Ath_CDS <- read_cds(file_path, format = "fasta")
 #' 
 #' }
-#' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
+#' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
 #' @export
 
 getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","CDS")){
