### 2017-03-15 282fb66
add functionality for ncrna retrieval
