### 2017-03-15 dcb4106
add functionality for ncrna retrieval
