+++ R/getGenome.R
@@ -39,7 +39,7 @@
 #' Ath_genome <- read_genome(file_path, format = "fasta")
 #' }
 #' 
-#' @seealso \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_genome}}
+#' @seealso \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_genome}}
 #' @export
 
 getGenome <-
