+++ R/getProteome.R
@@ -34,7 +34,7 @@
 #' 
 #' Ath_proteome <- read_proteome(file_path, format = "fasta")
 #' }
-#' @seealso \code{\link{getGenome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_proteome}}
+#' @seealso \code{\link{getGenome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_proteome}}
 #' @export
 
 getProteome <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","proteomes")){
