### 2017-03-15 75b8e89
new function for retrieving RNA files from NCBI and ENSEMBL databases
