### 2017-03-15 73b0ecd
add functionality for bulk RNA file retrieval
