### 2017-03-15 fb3ea3c
new function for importing RNA files that were retrieved with getRNA()
