+++ man/getCDS.Rd
@@ -35,7 +35,7 @@ Ath_CDS <- read_cds(file_path, format = "fasta")
 }
 }
 \seealso{
-\code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
+\code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
 }
 \author{
 Hajk-Georg Drost
