+++ man/getGenome.Rd
@@ -59,7 +59,7 @@ Ath_genome <- read_genome(file_path, format = "fasta")
 
 }
 \seealso{
-\code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_genome}}
+\code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_genome}}
 }
 \author{
 Hajk-Georg Drost
