+++ man/getProteome.Rd
@@ -54,7 +54,7 @@ Ath_proteome <- read_proteome(file_path, format = "fasta")
 }
 }
 \seealso{
-\code{\link{getGenome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_proteome}}
+\code{\link{getGenome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_proteome}}
 }
 \author{
 Hajk-Georg Drost
