+++ man/meta.retrieval.Rd
@@ -17,10 +17,14 @@ Available kingdoms can be retrieved with \code{\link{getKingdoms}}.}
 \item{db}{a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "emsembl"} or \code{db = "ensemblgenomes"}.}
 
 \item{type}{type of sequences that shall be retrieved. Options are:
-\code{type = "genome"} (for genome assembly retrieval; see also \code{\link{getGenome}}), \code{type = "proteome"} (for proteome retrieval; see also \code{\link{getProteome}}),
-\code{type = "CDS"} (for coding sequence retrieval; see also \code{\link{getCDS}}),
-\code{type = "gff"} (for annotation file retrieval in gff format; see also \code{\link{getGFF}}),
-\code{type = "assemblystats"} (for genome assembly quality stats file retrieval; see also \code{\link{getAssemblyStats}}).}
+\itemize{
+ \item \code{type = "genome"} (for genome assembly retrieval; see also \code{\link{getGenome}}), 
+ \item \code{type = "proteome"} (for proteome retrieval; see also \code{\link{getProteome}}),
+ \item \code{type = "CDS"} (for coding sequence retrieval; see also \code{\link{getCDS}}),
+ \item \code{type = "gff"} (for annotation file retrieval in gff format; see also \code{\link{getGFF}}),
+ \item \code{type = "rna"} (for RNA file retrieval in fasta format; see also \code{\link{getRNA}}),
+ \item \code{type = "assemblystats"} (for genome assembly quality stats file retrieval; see also \code{\link{getAssemblyStats}}).
+ }}
 
 \item{combine}{just in case \code{type = "assemblystats"} is specified, shall assemby stats of individual species be imported and combined to a \code{\link{data.frame}}?}
 
@@ -28,7 +32,7 @@ Available kingdoms can be retrieved with \code{\link{getKingdoms}}.}
 kingdom name is used to name the output folder.}
 }
 \description{
-Download genomes, proteomes, or CDS of all species within a kingdom of life.
+Download genomes, proteomes, cds, gff, rna, or assembly stats files of all species within a kingdom of life.
 }
 \details{
 This function aims to perform bulk retrieval of the genomes, proteomes, cds, etc. of species
@@ -49,7 +53,7 @@ meta.retrieval(kingdom = "vertebrate_mammalian", db = "genbank", type = "genome"
 # get all available kingdoms for ensemblgenomes
 getKingdoms(db = "ensemblgenomes")
 # download all vertebrate genomes from ensemblgenomes
-meta.retrieval(kingdom = "", db = "ensemblgenomes", type = "genome")
+meta.retrieval(kingdom = "vertebrate_mammalian", db = "ensemblgenomes", type = "genome")
 
 # In case users do not wish to retrieve genomes from an entire kingdom, 
 # but rather from a subgoup (e.g. from species belonging to the Gammaproteobacteria class,
