+++ man/meta.retrieval.all.Rd
@@ -10,13 +10,17 @@ meta.retrieval.all(db = "refseq", type = "genome")
 \item{db}{a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "emsembl"} or \code{db = "ensemblgenomes"}.}
 
 \item{type}{type of sequences that shall be retrieved. Options are:
-\code{type = "genome"} (for genome assembly retrieval; see also \code{\link{getGenome}}), \code{type = "proteome"} (for proteome retrieval; see also \code{\link{getProteome}}),
-\code{type = "CDS"} (for coding sequence retrieval; see also \code{\link{getCDS}}),
-\code{type = "gff"} (for annotation file retrieval in gff format; see also \code{\link{getGFF}}),
-\code{type = "assemblystats"} (for genome assembly quality stats file retrieval; see also \code{\link{getAssemblyStats}}).}
+\itemize{
+ \item \code{type = "genome"} (for genome assembly retrieval; see also \code{\link{getGenome}}), 
+ \item \code{type = "proteome"} (for proteome retrieval; see also \code{\link{getProteome}}),
+ \item \code{type = "CDS"} (for coding sequence retrieval; see also \code{\link{getCDS}}),
+ \item \code{type = "gff"} (for annotation file retrieval in gff format; see also \code{\link{getGFF}}),
+ \item \code{type = "rna"} (for RNA file retrieval in fasta format; see also \code{\link{getRNA}}),
+ \item \code{type = "assemblystats"} (for genome assembly quality stats file retrieval; see also \code{\link{getAssemblyStats}}).
+ }}
 }
 \description{
-Download genomes, proteomes, or CDS of individual species of all kingdoms of life.
+Download genomes, proteomes, cds, gff, rna, or assembly stats files of individual species of all kingdoms of life.
 }
 \details{
 This function aims to perform bulk retrieval of all genomes of species
