+++ man/read_proteome.Rd
@@ -40,7 +40,7 @@ Ath.proteome
 
 }
 \seealso{
-\code{\link{getProteome}}, \code{\link{read_genome}}, \code{\link{read_gff}}, \code{\link{read_cds}}
+\code{\link{getProteome}}, \code{\link{read_genome}}, \code{\link{read_gff}}, \code{\link{read_cds}}, \code{\link{read_rna}}
 }
 \author{
 Hajk-Georg Drost
