### 2017-03-15 2b63992
update unit test
