+++ tests/testthat/test-getCDS.R
@@ -30,7 +30,7 @@ test_that("The getCDS() interface works properly..",{
 test_that("The getCDS() error messages work properly..",{
         skip_on_cran()
         
-    expect_true(getCDS( db       = "ensembl",
+    expect_error(getCDS( db       = "ensembl",
                         organism = "Saccharomyces cerevisi",
                         path     = tempdir()))
 })
