### 2017-03-15 2e2c79a
add rna retrieval
