### 2017-03-15 1abd8e9
add rna retrieval options to meta retrieval vignette
