### 2017-03-15 70766f5
use RNAStringSet instead of BStringSet
