### 2017-03-15 fbd8bbc
add getRNA() example
