+++ vignettes/Sequence_Retrieval.Rmd
@@ -21,17 +21,19 @@ knitr::opts_chunk$set(
 
 The `biomartr` package allows users to retrieve biological sequences in a very simple and intuitive way.
 
-Using `biomartr`, users can retrieve either genomes, proteomes, or CDS data using the specialized functions:
+Using `biomartr`, users can retrieve either genomes, proteomes, CDS, RNA, GFF, and genome assembly statistics data using the specialized functions:
 
 * `getGenome()`
 * `getProteome()`
 * `getCDS()`
+* `getRNA()`
 * `getGFF()`
+* `getAssemblyStats()`
 
 
 ## Getting Started with Sequence Retrieval
 
-First users can check whether or not the genome, proteome, or CDS of their interest is available for download.
+First users can check whether or not the genome, proteome, CDS, RNA, GFF, or genome assembly statistics of their interest is available for download.
 
 Using the scientific name of the organism of interest, users can check whether the 
 corresponding genome is available via the `is.genome.available()` function.
@@ -1109,6 +1111,245 @@ Cress_CDS
 [35385]  1689 ATGAGTTTAACAAAGAAAGCAAGTG...GAACCACAGGCCGGTCTCCTTAGA AT2G26280.1 cds:k...
 [35386]  1362 ATGGGTGTTTCTCTTCTGAAACAAC...GGTTTGGCCAATAATGTTATCTAG AT4G32600.1 cds:k...
 ```
+
+
+### RNA Retrieval
+
+The `getRNA()` function is an interface function to the [NCBI RefSeq](http://www.ncbi.nlm.nih.gov/refseq/about/), [NCBI Genbank](http://www.ncbi.nlm.nih.gov/genbank/about/),
+[ENSEMBL](http://www.ensembl.org/index.html) and [ENSEMBLGENOMES](http://ensemblgenomes.org/) databases from
+which corresponding RNA files can be retrieved. It works analogous to `getGenome()`, `getProteome()`, and `getCDS()`. 
+
+The `db` argument specifies from which database proteomes in `*.fasta` file format shall be retrieved.
+
+Options are:
+
+- `db = "refseq"` for retrieval from [NCBI RefSeq](http://www.ncbi.nlm.nih.gov/refseq/about/)
+- `db = "genbank"` for retrieval from [NCBI Genbank](http://www.ncbi.nlm.nih.gov/genbank/about/)
+- `db = "ensembl"` for retrieval from [ENSEMBL](http://www.ensembl.org/index.html)
+- `db = "ensemblgenomes"` for retrieval from [ENSEMBLGENOMES](http://ensemblgenomes.org/)
+
+Furthermore, again users need to specify the scientific name of the organism of interest for which a proteomes shall be downloaded, e.g. `organism = "Homo sapiens"`.
+Finally, the `path` argument specifies the folder path in which the corresponding RNA file shall be locally stored. In case users would like to store the RNA file at a different location,
+they can specify the `path = file.path("put","your","path","here")` argument.
+
+Example `RefSeq`:
+
+```{r,eval=FALSE}
+# download the RNA of Homo sapiens from refseq
[…]
