### 2017-03-19 #93 @naupaka
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I should say right up here at the front that this is my first review for ROpenSci, so it probably took me a bit longer than others might have to get up to speed. It's also a relatively large package in terms of the number of functions it exposes to the user, and it took me a while to figure out how the parts fit together. I'm sure I missed things, but hopefully having me muddle through the codebase will provide some opportunities to smooth a few rough edges.

The goal of the package, as I understand it, is to provide a higher-level interface for programmatically downloading sequence or other datasets from major repositories. Overall, I think the package fills a useful niche and I plan to use it for my own research needs in the future. I hope that the commentary below is helpful.

I downloaded and read the [Bioinformatics paper that describes the package](https://doi.org/10.1093/bioinformatics/btw821). I found it helpful for understanding the goals of the package, and how it differs from biomaRt, which was something I was initially confused about.

I think it would be nice to clarify in the README and elsewhere that the queries work, for the most part (as far as I understand it), against NCBI or EMSEMBL, and not from any of the many other sources that have a BioMart interface (as @grimbough pointed out). For example, I am interested in getting data from the BioMart interface to [JGI's Phytozome](https://phytozome.jgi.doe.gov/biomart/martview/e8c2216498034ca35960fc6b993cd697), since they have a 3.0 version of the *Populus trichocarpa* genome (the NCBI version is still at 2.0), but that is not possible, I believe, with this package (*feature request* 😃).

Another thing that I didn't see, but I think would be useful, would be to modify `listGenomes()` to return a data frame where the first column is is the species' Latin binomial and the second column is which database each species' genome is associated with. Otherwise it's tricky to figure out which to set for the `db = ` parameter when using `getGenome()`.

### Minor comments and other issues

No Travis badge in README.

While all exported functions have roxygen-generated documentation, some internal functions have no comments or documentation, e.g. `connected.to.internet.R`, `exists.ftp.file.R`, or `get.emsembl.info.R`. You could add some short roxygen comments and add `#' @noRd` as recommended in the ROpenSci package guidelines, or just have some non-roxygen style comments in there in case others would like to better understand what's going on.

I also agree with @grimbough that it would be ideal if at least some of the example code in the documentation could actually be run. The output from `devtools::run_examples()` suggests that there isn't any code in any of the examples for any of the functions that is actually run. Obviously, it wouldn't make sense to run `download.database.all(name = "nr", path = "nr")` because that could take days, but surely there are some queries that would be fast enough to run in a reasonable timeframe?

##### Function/variable naming & general syntax

* I was a bit confused by which functions did what, in terms of which database(s) they query, and where exactly datasets were coming from. For each download, this information is all very helpfully documented in the `doc_*.txt` files in the `_ncbi_downloads` directory, but I felt like it was a bit of a hunt to figure out what had happened. Maybe this is just because I am still not 100% clear on which databases this package interfaces with. This is related to my earlier comment about being confused about how to figure out which db to get a genome from if you don't know ahead of time.

##### Suggestions

* [ROpenSci's packaging guide](https://github.com/ropensci/packaging_guide) suggests the use of `message()` and/or `warning()` instead of `print()` (e.g. lines 144 and 291 of `getProteome.R`).

* Suggestions from the `goodpractice` package:

```r
── GP biomartr ──────────────────────────────────────────────────────────────────────────

  ✖ write unit tests for all functions, and all package code in general. 0% of
    code lines are covered by test cases.

    R/biomart.R:49:NA
    R/biomart.R:50:NA
    R/biomart.R:51:NA
    R/biomart.R:53:NA
    R/biomart.R:54:NA

  ✖ write short and simple functions. These functions have high cyclomatic
    complexity:listGenomes (78), meta.retrieval (63).
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines

    R/biomart.R:71:1
    R/biomart.R:72:1
    R/connected.to.internet.R:5:1
    R/download_database.R:38:1
    R/download_database.R:52:1

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list,
    depending on the input data. Consider using vapply() instead.

    R/download_database_all.R:21:5
    R/download_database.R:31:35
    R/getMetaGenomeAnnotations.R:96:31
    R/getMetaGenomes.R:96:31
    R/listDatabases.R:78:29

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...)
    expressions. They are error prone and result 1:0 if the expression on the
    right hand side is zero. Use seq_len() or seq_along() instead.

    R/getAttributes.R:99:36
    R/getDatasets.R:55:39
    R/getFilters.R:99:40
    R/organismAttributes.R:79:39
    R/organismFilters.R:75:43

  ✖ not import packages as a whole, as this can cause name clashes between the
    imported packages. Instead, import only the specific functions you need.
  ✖ fix this R CMD check NOTE: Namespace in Imports field not imported from:
    ‘stringi’ All declared Imports should be used.
```

A number of the tests fail for me when run with either `R CMD CHECK` or `devtools::check()`. Not sure if it is appropriate to put all that output here, but I figured maybe it would be helpful for diagnosing the problem. I think these errors are related to the occasional 'weird server error' messages I was getting.

```r
==> devtools::check(document = FALSE)

Setting env vars ---------------------------------------------------------------
CFLAGS  : -Wall -pedantic
CXXFLAGS: -Wall -pedantic
Building biomartr --------------------------------------------------------------
'/usr/local/Cellar/r/3.3.3/R.framework/Resources/bin/R' --no-site-file  \
  --no-environ --no-save --no-restore --quiet CMD build  \
  '/Users/naupaka/git/ROpenSci/biomartr' --no-resave-data --no-manual

* checking for file ‘/Users/naupaka/git/ROpenSci/biomartr/DESCRIPTION’ ... OK
* preparing ‘biomartr’:
* creating vignettes ... OK
* checking for LF line-endings in source and make files
* checking for empty or unneeded directories
* building ‘biomartr_0.4.0.tar.gz’

Setting env vars ---------------------------------------------------------------
_R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
_R_CHECK_CRAN_INCOMING_           : FALSE
_R_CHECK_FORCE_SUGGESTS_          : FALSE
Checking biomartr --------------------------------------------------------------
'/usr/local/Cellar/r/3.3.3/R.framework/Resources/bin/R' --no-site-file  \
  --no-environ --no-save --no-restore --quiet CMD check  \
  '/var/folders/1y/8s_b2fx520sgj5f3j9fjwg880000gn/T//RtmpyLUWwj/biomartr_0.4.0.tar.gz'  \
  --as-cran --timings --no-manual

* using log directory ‘/Users/naupaka/git/ROpenSci/biomartr.Rcheck’
* using R version 3.3.3 (2017-03-06)
* using platform: x86_64-apple-darwin16.4.0 (64-bit)
* using session charset: UTF-8
* using options ‘--no-manual --as-cran’
* checking for file ‘biomartr/DESCRIPTION’ ... OK
* this is package ‘biomartr’ version ‘0.4.0’
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘biomartr’ can be installed ... OK
* checking installed package size ... OK
* checking package directory ... OK
* checking ‘build’ directory ... OK
* checking top-level files ... OK
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking R files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... OK
* checking whether the package can be loaded with stated dependencies ... OK
* checking whether the package can be unloaded cleanly ... OK
* checking whether the namespace can be loaded with stated dependencies ... OK
* checking whether the namespace can be unloaded cleanly ... OK
* checking loading without being on the library search path ... OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... OK
* checking Rd files ... OK
* checking Rd metadata ... OK
* checking Rd line widths ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
 WARNING
‘qpdf’ is needed for checks on size reduction of PDFs
* checking installed files from ‘inst/doc’ ... OK
* checking files in ‘vignettes’ ... OK
* checking examples ... OK
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
  Running ‘testthat.R’ [147s/511s]
 ERROR
  4: tryCatchOne(expr, names, parentenv, handlers[[1L]])
  5: value[[3L]](cond)
  6: stop(paste0("File ", file, " could not be read properly. \n", "Please make sure that ",
         file, " contains only amino acid sequences and is in ", format, " format."),
         call. = FALSE)

  testthat results ================================================================
  OK: 35 SKIPPED: 0 FAILED: 4
  1. Failure: The download.database() throws error when wrong input database is specified.. (@test-download.database.R#9)
  2. Error: The getGFF() interface works properly.. (@test-getGFF.R#31)
  3. Failure: The getGFF() error messages work properly.. (@test-getGFF.R#41)
  4. Error: The getProteome() interface works properly.. (@test-getProteome.R#24)

  Error: testthat unit tests failed
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes in ‘inst/doc’ ... OK
* checking re-building of vignette outputs ... OK
* DONE
Status: 1 ERROR, 1 WARNING

See
  ‘/Users/naupaka/git/ROpenSci/biomartr.Rcheck/00check.log’

R CMD check results
1 error  | 0 warnings | 0 notes
checking tests ... ERROR
  Running ‘testthat.R’ [147s/511s]
  4: tryCatchOne(expr, names, parentenv, handlers[[1L]])
  5: value[[3L]](cond)
  6: stop(paste0("File ", file, " could not be read properly. \n", "Please make sure that ",
         file, " contains only amino acid sequences and is in ", format, " format."),
         call. = FALSE)

  testthat results ================================================================
  OK: 35 SKIPPED: 0 FAILED: 4
  1. Failure: The download.database() throws error when wrong input database is specified.. (@test-download.database.R#9)
  2. Error: The getGFF() interface works properly.. (@test-getGFF.R#31)
  3. Failure: The getGFF() error messages work properly.. (@test-getGFF.R#41)
  4. Error: The getProteome() interface works properly.. (@test-getProteome.R#24)

  Error: testthat unit tests failed

R CMD check succeeded
```

#### Minor documentation issues

Looks like there is a non-ASCII character in the copied tibble output in one of the verbatim blocks in the Functional Annotation vignette (non-executing).

```r
==> devtools::document(roclets=c('rd', 'collate', 'namespace', 'vignette'))

Updating biomartr documentation
Loading biomartr
Updating vignettes
Rebuilding BioMart_Examples.Rmd
Rebuilding Database_Retrieval.Rmd
Rebuilding Functional_Annotation.Rmd
Error in tools::buildVignette(vign, dirname(vign), tangle = FALSE, clean = FALSE) :
  Vignette 'Functional_Annotation' is non-ASCII but has no declared encoding
Calls: suppressPackageStartupMessages ... roclet_output.roclet_vignette -> vign_update_all -> vapply -> FUN -> <Anonymous>

Exited with status 1.
```

```r
> tools::showNonASCII(readLines("vignettes/Functional_Annotation.Rmd"))
# 566: A tibble: 6 <c3><97> 4
```

#### Repeated "No encoding supplied" warnings

```r
> organismBM(organism = "Homo sapiens")
No encoding supplied: defaulting to UTF-8.
No encoding supplied: defaulting to UTF-8.
No encoding supplied: defaulting to UTF-8.
No encoding supplied: defaulting to UTF-8.
# A tibble: 12 × 5
   organism_name
           <chr>
1       hsapiens
2       hsapiens
3       hsapiens
4       hsapiens
5       hsapiens
6       hsapiens
7       hsapiens
8       hsapiens
9       hsapiens
10      hsapiens
11      hsapiens
12      hsapiens
# ... with 4 more variables: description <chr>, mart <chr>, dataset <chr>,
#   version <chr>
```

#### Error when downloading assembly stats

Sometimes when I would try to run a command, I'd get errors like this. I was connected to the Internet at the time, and similar commands run before and afterwards worked fine. Sometimes the same command, run a second time, worked fine even though it failed at first. Maybe something about the timeout before waiting for a server response could be tweaked to reduce the likelihood of premature failure?

```r
> meta.retrieval(kingdom = "plant",
               db      = "refseq",
               type    = "assemblystats",
               combine = TRUE)
Error: The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site 'ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/001/190/045/GCF_001190045.1_Vigan1.1/GCF_001190045.1_Vigan1.1_assembly_stats.txt' currently available?
In download.file(url, ...) :
  URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/001/190/045/GCF_001190045.1_Vigan1.1/GCF_001190045.1_Vigan1.1_assembly_stats.txt': status was 'Weird server reply'
```

### Session Info

Tested with commit 2570b0d3296 on master branch from https://github.com/HajkD/biomartr

```r
> sessionInfo()
Platform: x86_64-apple-darwin16.4.0 (64-bit)
Running under: macOS Sierra 10.12.3

[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

[1] biomartr_0.4.0

 [1] Rcpp_0.12.9          IRanges_2.8.1        XML_3.98-1.5         Biostrings_2.42.1    digest_0.6.12
 [6] bitops_1.0-6         DBI_0.6              stats4_3.3.3         RSQLite_1.1-2        zlibbioc_1.20.0
[11] curl_2.3             XVector_0.14.0       S4Vectors_0.12.1     tools_3.3.3          Biobase_2.34.0
[16] biomaRt_2.30.0       RCurl_1.95-4.8       parallel_3.3.3       BiocGenerics_0.20.0  AnnotationDbi_1.36.2
[21] memoise_1.0.0
```
