### 2017-03-21 #93 @HajkD
Dear @sckott,

Thank you very much for retrieving two valuable and in-depth reviews from @grimbough and @naupaka for the `biomartr` package.

I am very happy to see that both reviewers appreciate the functionality of `biomartr`. The weaknesses and potential improvements of the package that were pointed out by both reviewers are very constructive and helped me a lot to spot existing misunderstandings and problems.

I already started to incorporate the suggestions made by @grimbough and will now start to address all issues raised by @naupaka.

I highly appreciate the time the reviewers took to generate an in-depth code review and I will now work on a detailed response letter.

Kind regards,
Hajk
