### 2017-03-21 369d155
new helper function to assure a stable download
