### 2017-03-21 9378ba2
new functionality: file is not downloaded again if it already exists in the specified path
