### 2017-03-21 5f8f5de
new functionality: file is not downloaded again if it already exists in the specified path
