+++ R/getCDS.R
@@ -86,23 +86,35 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             }
             
             if (nrow(FoundOrganism) == 1) {
-                tryCatch({
-                    utils::capture.output(downloader::download(
-                        download_url,
-                        destfile = file.path(
-                            path,
-                            paste0(local.org, "_cds_from_genomic_", db, ".fna.gz")
-                        ),
-                        mode = "wb"
-                    ))
-                }, error = function(e)
-                    stop(
-                        "The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site '",
-                        download_url,
-                        "' currently available?",
-                        call. = FALSE
-                    ))
-                
+                if (file.exists(file.path(
+                        path,
+                        paste0(local.org, "_cds_from_genomic_", db, ".fna.gz")
+                ))) {
+                        message("File ",
+                                file.path(
+                                        path,
+                                        paste0(local.org, "_cds_from_genomic_", db, ".fna.gz")
+                                ),
+                                " exists already. Thus, download has been skipped.")
+                } else {
+                        tryCatch({
+                                utils::capture.output(custom_download(
+                                        download_url,
+                                        destfile = file.path(
+                                                path,
+                                                paste0(local.org, "_cds_from_genomic_", db, ".fna.gz")
+                                        ),
+                                        mode = "wb"
+                                ))
+                        }, error = function(e)
+                                stop(
+                                        "The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site '",
+                                        download_url,
+                                        "' currently available?",
+                                        call. = FALSE
+                                ))  
+                }   
+                    
                 docFile(
                     file.name = paste0(local.org, "_cds_from_genomic_", db, ".fna.gz"),
                     organism  = organism,
@@ -122,10 +134,7 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
                     submitter = FoundOrganism$submitter
                 )
                 
-                # NCBI limits requests to three per second
-                Sys.sleep(0.33)
-                
-                print(
+                message(
                     paste0(
                         "The genomic CDS of '",
                         organism,
@@ -204,7 +213,7 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             
             sink()
             
-            print(
+            message(
                 paste0(
                     "The CDS of '",
                     organism,
@@ -274,7 +283,7 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             
             sink()
         
-            print(
+            message(
                 paste0(
                     "The CDS of '",
                     organism,
