### 2017-03-21 0f6deda
new functionality: file is not downloaded again if it already exists in the specified path
