+++ R/getENSEMBL.Annotation.R
@@ -94,25 +94,53 @@ getENSEMBL.Annotation <- function(organism, type = "dna", id.type = "toplevel", 
             )
         )
     
-    tryCatch({
-        downloader::download(ensembl.qry,
-                             destfile = file.path(
-                                 path,
-                                 paste0(
-                                     new.organism,
-                                     ".",
-                                     json.qry.info$default_coord_system_version,
-                                     ".",
-                                     ensembl.available.organisms$release[1],
-                                     "_ensembl",
-                                     ".gff3.gz"
-                                 )
-                             ),
-                             mode = "wb")
-    }, error = function(e)
-        stop(
-            "The FTP site of ENSEMBL 'ftp://ftp.ensembl.org/pub/current_gff3/' does not seem to work properly. Are you connected to the internet? Is the site 'ftp://ftp.ensembl.org/pub/current_gff3/' or 'http://rest.ensembl.org' currently available?", call. = FALSE
-        ))
+    if (file.exists(file.path(
+            path,
+            paste0(
+                    new.organism,
+                    ".",
+                    json.qry.info$default_coord_system_version,
+                    ".",
+                    ensembl.available.organisms$release[1],
+                    "_ensembl",
+                    ".gff3.gz"
+            )
+    ))) {
+            
+            message("File ",file.path(
+                    path,
+                    paste0(
+                            new.organism,
+                            ".",
+                            json.qry.info$default_coord_system_version,
+                            ".",
+                            ensembl.available.organisms$release[1],
+                            "_ensembl",
+                            ".gff3.gz"
+                    )
+            )," exists already. Thus, download has been skipped.")
+    } else {
+            tryCatch({
+                    custom_download(ensembl.qry,
+                                    destfile = file.path(
+                                            path,
+                                            paste0(
+                                                    new.organism,
+                                                    ".",
+                                                    json.qry.info$default_coord_system_version,
+                                                    ".",
+                                                    ensembl.available.organisms$release[1],
+                                                    "_ensembl",
+                                                    ".gff3.gz"
+                                            )
+                                    ),
+                                    mode = "wb")
+            }, error = function(e)
+                    stop(
+                            "The FTP site of ENSEMBL 'ftp://ftp.ensembl.org/pub/current_gff3/' does not seem to work properly. Are you connected to the internet? Is the site 'ftp://ftp.ensembl.org/pub/current_gff3/' or 'http://rest.ensembl.org' currently available?", call. = FALSE
+                    ))
+    }
+    
     
     return(file.path(
         path,
