### 2017-03-21 bddfd98
new functionality: file is not downloaded again if it already exists in the specified path
