### 2017-03-21 ee595c8
new functionality: file is not downloaded again if it already exists in the specified path
