### 2017-03-21 89eaef9
new functionality: file is not downloaded again if it already exists in the specified path
