+++ R/getGFF.R
@@ -111,15 +111,25 @@ getGFF <-
                 local.org <- stringr::str_replace_all(organism,"\\/","_")
                 
                 if (nrow(FoundOrganism) == 1) {
-                    tryCatch({utils::capture.output(downloader::download(
-                        download_url,
-                        destfile = file.path(path, paste0(local.org, "_genomic_",db,".gff.gz")),
-                        mode = "wb"
-                    ))}, error = function(e)
-                        stop(
-                            "The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site '",download_url,"' currently available?", call. = FALSE
-                        ))
-                    
+                        if (file.exists(file.path(path, 
+                                paste0(local.org, "_genomic_", db, ".gff.gz")))) {
+                                
+                                message("File ",
+                                        file.path(path, 
+                                                  paste0(local.org, "_genomic_", db, ".gff.gz")),
+                                        " exists already. Thus, download has been skipped.")
+                        } else {
+                                tryCatch({utils::capture.output(custom_download(
+                                        download_url,
+                                        destfile = file.path(path, paste0(local.org, "_genomic_",db,".gff.gz")),
+                                        mode = "wb"
+                                ))}, error = function(e)
+                                        stop(
+                                                "The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site '",download_url,"' currently available?", call. = FALSE
+                                        ))
+                        }
+                        
+        
                     docFile(
                         file.name = paste0(local.org, "_genomic_",db,".gff.gz"),
                         organism  = organism,
@@ -139,11 +149,7 @@ getGFF <-
                         submitter = FoundOrganism$submitter
                     )
                     
-                    # NCBI limits requests to three per second
-                    Sys.sleep(0.33)
-                    
-                    
-                    print(
+                    message(
                         paste0(
                             "The *.gff annotation file of '",
                             organism,
@@ -219,7 +225,7 @@ getGFF <-
                 
                 sink()
                 
-                print(
+                message(
                     paste0(
                         "The *.gff annotation file of '",
                         organism,
@@ -288,7 +294,7 @@ getGFF <-
                 
                 sink()
             
-                print(
+                message(
                     paste0(
                         "The *.gff annotation file of '",
                         organism,
