### 2017-03-21 c3d8c17
new functionality: file is not downloaded again if it already exists in the specified path
