### 2017-03-21 dab57be
new functionality: file is not downloaded again if it already exists in the specified path
