### 2017-03-21 5cb39d4
new functionality: file is not downloaded again if it already exists in the specified path
