+++ R/getRNA.R
@@ -86,26 +86,38 @@ getRNA <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             }
             
             if (nrow(FoundOrganism) == 1) {
-                tryCatch({
-                    utils::capture.output(downloader::download(
-                        download_url,
-                        destfile = file.path(
-                            path,
-                            paste0(local.org, "_rna_from_genomic_", db, ".fna.gz")
-                        ),
-                        mode = "wb"
-                    ))
-                }, error = function(e) {
-                    warning(
-                        "The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site '",
-                        download_url,
-                        "' currently available?",
-                        call. = FALSE
-                    )
-                    return(FALSE)
-                }
-                    )
-                
+                if (file.exists(file.path(
+                        path,
+                        paste0(local.org, "_rna_from_genomic_", db, ".fna.gz")
+                ))) {
+                        message("File ",
+                                file.path(
+                                        path,
+                                        paste0(local.org, "_rna_from_genomic_", db, ".fna.gz")
+                                ),
+                                " exists already. Thus, download has been skipped.")
+                } else {
+                        tryCatch({
+                                utils::capture.output(custom_download(
+                                        download_url,
+                                        destfile = file.path(
+                                                path,
+                                                paste0(local.org, "_rna_from_genomic_", db, ".fna.gz")
+                                        ),
+                                        mode = "wb"
+                                ))
+                        }, error = function(e) {
+                                warning(
+                                        "The FTP site 'ftp://ftp.ncbi.nlm.nih.gov/' cannot be reached. Are you connected to the internet? Is the the FTP site '",
+                                        download_url,
+                                        "' currently available?",
+                                        call. = FALSE
+                                )
+                                return(FALSE)
+                        }
+                        )
+                }   
+                  
                 docFile(
                     file.name = paste0(local.org, "_rna_from_genomic_", db, ".fna.gz"),
                     organism  = organism,
@@ -125,10 +137,7 @@ getRNA <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
                     submitter = FoundOrganism$submitter
                 )
                 
-                # NCBI limits requests to three per second
-                Sys.sleep(0.33)
-                
-                print(
+                message(
                     paste0(
                         "The genomic RNA of '",
                         organism,
@@ -210,7 +219,7 @@ getRNA <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             
             sink()
             
-            print(
+            message(
                 paste0(
                     "The RNA of '",
                     organism,
@@ -282,7 +291,7 @@ getRNA <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             
             sink()
             
-            print(
+            message(
                 paste0(
                     "The RNA of '",
                     organism,
