### 2017-03-21 dd2b5e5
update NEWS
