### 2017-03-22 086aa0a
modifying error handling
