+++ R/getENSEMBL.Annotation.R
@@ -9,8 +9,8 @@ getENSEMBL.Annotation <- function(organism, type = "dna", id.type = "toplevel", 
     new.organism <- stringr::str_replace_all(organism, " ", "_")
     
     # if (file.exists(file.path(tempdir(), "ensembl_summary.txt"))) {
-    #     suppressWarnings(ensembl.available.organisms <-
-    #                          readr::read_tsv(
+    #     ensembl.available.organisms <-
+    #                          readr::read_delim(
     #                              file.path(tempdir(), "ensembl_summary.txt"),
     #                              col_names = c(
     #                                  "division",
@@ -32,27 +32,28 @@ getENSEMBL.Annotation <- function(organism, type = "dna", id.type = "toplevel", 
     #                                  common_name = readr::col_character(),
     #                                  assembly = readr::col_character()
     #                              ),
-    #                              comment = "#"
-    #                          ))
+    #                              comment = "#",
+    #                              delim = "\t"
+    #                          )
     # }
-    
+    # 
     if (!file.exists(file.path(tempdir(), "ensembl_summary.txt"))) {
-        # check if organism is available on ENSEMBL
-        tryCatch({
+       # check if organism is available on ENSEMBL
+       tryCatch({
             ensembl.available.organisms <-
                 jsonlite::fromJSON("http://rest.ensembl.org/info/species?content-type=application/json")
         }, error = function(e)
             stop(
                 "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensembl.org' currently available?", call. = FALSE
             ))
-        
+
         aliases <- groups <- NULL
         # transform list object returned by 'fromJSON' to tibble
         ensembl.available.organisms <-
             tibble::as_tibble(dplyr::select(ensembl.available.organisms$species, -aliases, -groups))
         
-        # readr::write_tsv(ensembl.available.organisms,
-        #                  file.path(tempdir(), "ensembl_summary.txt"), col_names = TRUE)
+        # readr::write_delim(ensembl.available.organisms,
+        #                  file.path(tempdir(), "ensembl_summary.txt"), col_names = TRUE, delim = "\t")
     }
     
     if (!is.element(stringr::str_to_lower(new.organism),
