### 2017-03-22 58a6093
add error handling
