+++ R/exists.ftp.file.R
@@ -1,9 +1,13 @@
 #' @import curl
 exists.ftp.file <- function(url, file.path) {
-    con <- curl::curl(url = url, "rb", curl::new_handle(dirlistonly = TRUE))
+        tryCatch({
+                con <-
+                        curl::curl(url = url,
+                                   open = "rb",
+                                   curl::new_handle(dirlistonly = TRUE))
+        }, error = function(e) stop("Something went wrong with the connection to: ",url)) 
     ftp.content <- suppressMessages(readr::read_delim(con, delim = "\n", col_names = FALSE))
     close(con)
-    Sys.sleep(0.33)
     if (is.element(as.character(basename(file.path)), as.character(ftp.content$X1)))
         return(TRUE)
     else 
