### 2017-03-22 a0f6a04
fix bug https://github.com/HajkD/biomartr/issues/7 -> the problem was that ENSEMBLGENOMES stores bacteria genomes in so called "collections"; when checking the file availability with exists.ftp.file() this collection structure was not considered
