+++ R/getENSEMBL.Seq.R
@@ -111,12 +111,7 @@ getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
             )
         )
     
-    if (!exists.ftp.file(url = paste0(
-        "ftp://ftp.ensembl.org/pub/current_fasta/",
-        stringr::str_to_lower(new.organism),
-        "/",
-        type,
-        "/"), file.path = ensembl.qry)) {
+    if (!exists.ftp.file(url = ensembl.qry, file.path = ensembl.qry)) {
         message("Unfortunately no ",type," file could be found for organism '",organism,"'. Thus, the download of this organism has been omitted.")
         return(FALSE) 
     }
