### 2017-03-22 7e20593
update NEWS
