+++ NEWS.md
@@ -1,6 +1,10 @@
 biomartr 0.5.0
 ===========
 
+### Bug fixes
+
+- fixing a bug in `exists.ftp.file()` and `getENSEMBLGENOMES.Seq()` that caused bacterial genome, proteome, etc retrieval to fail due to the wrong construction of a query ftp request https://github.com/HajkD/biomartr/issues/7
+
 ### New Functions
 
 - new function `getRNA()` to perform RNA Sequence Retrieval from NCBI and ENSEMBL databases (Thanks for suggesting it @carlo-berg)
