+++ NEWS.md
@@ -4,6 +4,7 @@ biomartr 0.5.0
 ### Bug fixes
 
 - fixing a bug in `exists.ftp.file()` and `getENSEMBLGENOMES.Seq()` that caused bacterial genome, proteome, etc retrieval to fail due to the wrong construction of a query ftp request https://github.com/HajkD/biomartr/issues/7
+(Many thanks to @dbsseven)
 
 ### New Functions
 
