### 2017-03-22 69c66a2
now partial string matching is used to search for organism names.. allowing a more variable search for organisms. see https://github.com/HajkD/biomartr/issues/8
