### 2017-03-22 a13090d
adding partial string matching also for ensembl
