+++ vignettes/Functional_Annotation.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Functional Annotation with BioMart, GO, and KeGG"
+title: "Functional Annotation with Ensembl Biomart, GO, and KeGG"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
@@ -17,29 +17,31 @@ knitr::opts_chunk$set(
 ```
 
 
-## Functional Annotation with BioMart
+## Functional Annotation with `Ensembl Biomart`
 
-The [BioMart](http://biomart.org/) project enables users to retrieve a vast diversity of annotation data
-for specific organisms. Steffen Durinck and Wolfgang Huber provide an powerful interface between
-the R language and [BioMart](http://biomart.org/) by providing the R package [biomaRt](http://www.bioconductor.org/packages/release/bioc/html/biomaRt.html).
-The following sections will introduce users to the functionality and data retrieval precedures using the `biomaRt`
-package and will then introduce them to the interface functions `biomart()` and `biomart_organisms()` implemented in `biomartr`
-that are based on the `biomaRt` methodology but aim to introduce an more intuitive way of interacting with BioMart.
+The [Ensembl Biomart](http://ensemblgenomes.org/info/access/biomart) database enables users to retrieve a vast diversity of annotation data
+for specific organisms. Initially, Steffen Durinck and Wolfgang Huber provide an powerful interface between
+the R language and [Ensembl Biomart](http://ensemblgenomes.org/info/access/biomart) by providing the R package [biomaRt](http://www.bioconductor.org/packages/release/bioc/html/biomaRt.html).
+However, the `biomartr` package extends the functionality of the [biomaRt](http://www.bioconductor.org/packages/release/bioc/html/biomaRt.html) package
+and introduces a more organism centered annotation retrieval concept.
+
+The following sections will introduce users to the functionality and data retrieval precedures of `biomartr` and will show how `biomartr`
+extends the functionality of the initial [biomaRt](http://www.bioconductor.org/packages/release/bioc/html/biomaRt.html) package.
 
 ### Getting Started with biomaRt
 
 The best way to get started with the methodology presented by the established [biomaRt](http://www.bioconductor.org/packages/release/bioc/html/biomaRt.html) package is to understand the workflow
-of data retrieval. The database provided by [BioMart](http://biomart.org/) is organized in so called:
+of data retrieval. The database provided by [Ensembl Biomart](http://ensemblgenomes.org/info/access/biomart) is organized in so called:
 `marts`, `datasets`, and `attributes`. So when users want to retrieve information for a specific organism
 of interest, first they need to specify the `marts` and `datasets` in which the information of the corresponding
 organism can be found. Subsequently they can specify the `attributes` argument that is ought to be returned for the corresponding organism.
 
 The availability of `marts`, `datasets`, and `attributes` can be checked by the following functions:
 
-```{r,eval=FALSE}
+```{r,eval=TRUE}
 # install the biomaRt package
-source("http://bioconductor.org/biocLite.R")
-biocLite("biomaRt")
+# source("http://bioconductor.org/biocLite.R")
+# biocLite("biomaRt")
 
 # load biomaRt
 library(biomaRt)
@@ -48,37 +50,14 @@ library(biomaRt)
 head(listMarts(host = "www.ensembl.org"), 10)
 ```
 
-```
-               biomart               version
-1 ENSEMBL_MART_ENSEMBL      Ensembl Genes 83
-2     ENSEMBL_MART_SNP  Ensembl Variation 83
-3 ENSEMBL_MART_FUNCGEN Ensembl Regulation 83
-4    ENSEMBL_MART_VEGA               Vega 63
[…]
