### 2017-03-30 c319c24
update download procedure
