### 2017-03-31 HajkD/biomartr#9 @HajkD
Interface functions for JGI's Phytozome
Implement retrieval functions for JGI's Phytozome.
