+++ NAMESPACE
@@ -30,6 +30,7 @@ export(listGenomes)
 export(listGroups)
 export(listKingdoms)
 export(listMetaGenomes)
+export(listNCBIDatabases)
 export(meta.retrieval)
 export(meta.retrieval.all)
 export(organismAttributes)
