### 2017-04-01 17bd495
update NEWS
