### 2017-04-01 397880f
switch from listDatabases() to listNCBIDatabases()
