### 2017-04-01 0954528
now using listNCBIDatabases()
