+++ R/download_database_all.R
@@ -9,7 +9,7 @@
 #' @examples 
 #' \dontrun{
 #' # search for available NCBI databases
-#'   listDatabases(db = "all")
+#'   listNCBIDatabases(db = "all")
 #'   
 #' # choose database NCBI nr and download compelete database
 #'   download.database.all(name = "nr", path = "nr")
