### 2017-04-01 974d18d
rename listDatabases() to listNCBIDatabases()
