+++ R/listNCBIDatabases.R
@@ -1,0 +1,3 @@
+#' @export
+#' @rdname listDatabases
+listNCBIDatabases <- listDatabases
