+++ man/listDatabases.Rd
@@ -1,10 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/listDatabases.R
+% Please edit documentation in R/listDatabases.R, R/listNCBIDatabases.R
 \name{listDatabases}
 \alias{listDatabases}
+\alias{listNCBIDatabases}
 \title{Retrieve a List of Available NCBI Databases for Download}
 \usage{
 listDatabases(db = "nr", update = FALSE)
+
+listNCBIDatabases(db = "nr", update = FALSE)
 }
 \arguments{
 \item{db}{a character string specifying the name of the database that shall be searched for.}
@@ -24,13 +27,13 @@ select an appropriate database for download with \code{\link{download.database}}
 \examples{
 
 # retrieve all versions of the NCBI 'nr' database that can be downloaded
-# listDatabases(db = "nr")
+# listNCBIDatabases(db = "nr")
 
 # analogous:
-# listDatabases(db = "cdd")
-# listDatabases(db = "nt")
-# listDatabases(db = "gss")
-# listDatabases(db = "refseq_protein")
+# listNCBIDatabases(db = "cdd")
+# listNCBIDatabases(db = "nt")
+# listNCBIDatabases(db = "gss")
+# listNCBIDatabases(db = "refseq_protein")
 
 }
 \seealso{
