+++ R/getENSEMBL.Seq.R
@@ -1,3 +1,12 @@
+#' @title Helper function for retrieving biological sequence files from ENSEMBL
+#' @description This function downloads gff files of query organisms from ENSEMBL.
+#' @param organism scientific name of the organism of interest.
+#' @param type biological sequence type.
+#' @param id.type id type.
+#' @param path location where file shall be stored.
+#' @author Hajk-Georg Drost
+#' @noRd
+
 getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
     
     if (!is.element(type, c("dna", "cds", "pep", "ncrna")))
