+++ R/is.ensembl.alive.R
@@ -1,3 +1,8 @@
+#' @title Helper function to test if ENSEMBL server is reachable
+#' @description This function pings the ENSEMBL FTP server to 
+#' test whether or not a connection can be established.
+#' @author Hajk-Georg Drost
+#' @noRd
 is.ensembl.alive <- function() {
     
     ping <- jsonlite::fromJSON("http://rest.ensembl.org/info/ping?content-type=application/json")
