### 2017-04-02 HajkD/biomartr#10 @HajkD
Extend listGenomes() output
Modify listGenomes() to return a data frame where the first column is is the species' Latin binomial and the second column denotes the database each species' genome is associated with.
