### 2017-04-02 ff4c48e
update query structure
