### 2017-04-02 0db3a6d
update Intro vignette
