+++ tests/testthat/test-getGFF.R
@@ -1,32 +1,23 @@
 context("Test: getGFF()")
 
-test_that("The getGFF() interface works properly..",{
+test_that("The getGFF() interface works properly for RefSeq..",{
     
     skip_on_cran()
-    skip_on_travis() 
-    
+
     # test proper download from refseq
     getGFF( db       = "refseq", 
                    organism = "Homo sapiens", 
                    path = file.path("_ncbi_downloads","annotation"))
-    
-    # test proper use of internal referece files when command is repeated
-    
+})    
+
+test_that("The getGFF() interface works properly for Genbank..",{
     # test proper download from genbank
     getGFF( db       = "genbank", 
                    organism = "Homo sapiens", 
                    path = file.path("_ncbi_downloads","annotation"))
-    
-    # test proper use of internal referece files when command is repeated
-    getGFF( db       = "refseq", 
-                   organism = "Homo sapiens", 
-                   path = file.path("_ncbi_downloads","annotation"))
-    
-    # test proper download from genbank
-    getGFF( db       = "genbank", 
-                   organism = "Homo sapiens", 
-                   path = file.path("_ncbi_downloads","annotation"))
-    
+})    
+
+test_that("The getGFF() interface works properly for ENSEMBL",{
     # test proper download from ENSEMBL
     getGFF( db       = "ensembl", 
                    organism = "Saccharomyces cerevisiae", 
@@ -35,9 +26,3 @@ test_that("The getGFF() interface works properly..",{
 })
 
 
-test_that("The getGFF() error messages work properly..",{
-    skip_on_cran()
-    expect_output(getGFF( db       = "ensembl",
-                            organism = "Saccharomyces cerevisi",
-                            path     = tempdir()))
-})
