### 2017-04-02 a589883
update unit test
