+++ tests/testthat/test-getGFF.R
@@ -17,12 +17,3 @@ test_that("The getGFF() interface works properly for Genbank..",{
                    path = file.path("_ncbi_downloads","annotation"))
 })    
 
-test_that("The getGFF() interface works properly for ENSEMBL",{
-    # test proper download from ENSEMBL
-    getGFF( db       = "ensembl", 
-                   organism = "Saccharomyces cerevisiae", 
-                   path = file.path("_ncbi_downloads","annotation"))
-  
-})
-
-
