### 2017-04-03 #93 @sckott
@HajkD thanks for your changes and detailed comments

- For running egs, I suggest MAYBE running egs wrapped in dontrun on Travis by adding ` --run-dontrun` to your check args (since taking a longish time on travis I assume is not a big deal) - I'd have one or two examples for each function not wrapped in dontrun when you know they will run quickly. Note: I said MAYBE since check already takes a very long time, but still would be nice to make sure egs always work
- output of `goodpractice::gp()` was pointed out and there's some things in there which I think you should address:
    - `sapply` usage: replace with `vapply`/`lapply`
    -  it's a bit of work, but I highly suggest tidying code and documentation to 80 line width so you don't get text running off the page on the CRAN manual (e.g., https://cran.rstudio.com/web/packages/biomartr/biomartr.pdf), and it's easier to see in a narrow width editor setup
    -  replace usage of `1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...)` with eg., `seq_along`/`seq_len`
- Add a package level manual file - so that users can do `?biomartr` to get to the man pages easily
- looks like test coverage is about 47% - would be good to get that % up - not necessary now, but in the future.
- R cmd check takes a long time, just for ease of quick turn around time, it'd be wise to make that time much less - I assume most of the time is in tests, but haven't looked into how to speed that up
- also related to tests, but not about speed: looks like many of your test files aren't doing much testing using `testthat` - that is, you run examples, but then don't test their output (e.g., https://github.com/HajkD/biomartr/blob/master/tests/testthat/test-is.genome.available.R#L12-L36)  I imagine the majority of time in tests right now is doing data requests - adding test expectations shouldn't add a lot more time
