### 2017-04-04 #93 @HajkD
Dear @sckott,

Thank you so much for your suggestions. I will address all comments and will come back to you as soon as I finished.

Best wishes,
Hajk
