### 2017-04-06 f504dee
update NEWS
