### 2017-04-06 f7c4b6b
add md5 check sum test
