### 2017-04-06 95751fa
add md5 check sum test
