+++ R/download_database.R
@@ -1,79 +1,91 @@
 #' @title Download a NCBI Database to Your Local Hard Drive
 #' @description This function allows users to download a database selected by
 #' \code{\link{listDatabases}} to their local hard drive.
-#' @param db a character string specifying the database that shall be downloaded (selected from \code{\link{listDatabases}}).
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' database shall be stored. Default is \code{path = "database"}. In case this folder does not exist yet, it will be created.
+#' @param db a character string specifying the database that shall be downloaded
+#'  (selected from \code{\link{listDatabases}}).
+#' @param path a character string specifying the location (a folder) in
+#' which the corresponding database shall be stored.
+#' Default is \code{path = "database"}.
+#' In case this folder does not exist yet, it will be created.
 #' @details
-#' This function downloads large databases to your hard drive. For this purpose a folder
-#' named \code{database} (default) is created and the correspondning database then stored in this folder. 
-#' 
+#' This function downloads large databases to your hard drive.
+#' For this purpose a folder
+#' named \code{database} (default) is created and the correspondning
+#' database then stored in this folder.
 #' @author Hajk-Georg Drost
-#' @examples 
-#' \dontrun{
+#' @examples
 #'   # search for available NCBI nr databases
 #'   listNCBIDatabases(db = "nr")
-#'   
+#' \dontrun{
 #'   # select NCBI nr version 27 =  "nr.27.tar.gz"
 #'   # and download it to your hard drive
 #'   # -> please note that large databases take some time for download!
 #'   download.database(db = "nr.27.tar.gz")
-#' 
 #' }
 #' @seealso \code{\link{download.database.all}}, \code{\link{listDatabases}}
 #' @export
 
-download.database <- function(db, path = "database"){
-    
+download.database <- function(db, path = "database") {
     # test if internet connection is available
     connected.to.internet()
     
-    message("Starting download process of file: ",db, " ...")        
+    message("Starting download process of file: ", db, " ...")
     db.name <- names(table(unlist(sapply(db, function(x)
         unlist(stringr::str_split(x, "[.]"))[1]))))
     
-    if (!is.element(db.name, listNCBIDatabases("all")))
+    if (!is.element(db.name, listNCBIDatabases(db = "all")))
         stop(
             "The specified database '",
             db,
-            "' could not be found on NCBI. Please use the listNCBIDatabases() command to retrieve available databases or check if the name was written correctly.",
+            "' could not be found on NCBI.
+            Please use the listNCBIDatabases() command to retrieve available 
+            databases or check if the name was written correctly.",
             call. = FALSE
         )
-
+    
     if (!file.exists(path))
         dir.create(path)
     
     tryCatch({
-        custom_download(paste0("ftp://ftp.ncbi.nlm.nih.gov/blast/db/", db),
-                             file.path(path, db),
-                             mode = "wb")
-         
-        custom_download(paste0("ftp://ftp.ncbi.nlm.nih.gov/blast/db/", paste0(db,".md5")),
-                            file.path(path, paste0(db,".md5")),
-                            mode = "wb")    
+        custom_download(
+            paste0("ftp://ftp.ncbi.nlm.nih.gov/blast/db/", db),
+            file.path(path, db),
+            mode = "wb"
+        )
+        
+        custom_download(
+            paste0(
+                "ftp://ftp.ncbi.nlm.nih.gov/blast/db/",
+                paste0(db, ".md5")
+            ),
+            file.path(path, paste0(db, ".md5")),
+            mode = "wb"
+        )
     }, error = function(e)
         stop(
[…]
