### 2017-04-06 ab83746
format to 80 line and update examples
