+++ R/getAssemblyStats.R
@@ -1,12 +1,19 @@
 #' @title Genome Assembly Stats Retrieval
-#' @description  Main genome assembly stats retrieval function for an organism of interest.
-#' By specifying the scientific name of an organism of interest the corresponding  genome assembly stats file storing the assembly statistics of the organism of interest
-#' can be downloaded and stored locally. Genome assembly stats files can be retrieved from several databases.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"} or \code{db = "genbank"}.
-#' @param organism a character string specifying the scientific name of the organism of interest, e.g. \code{organism = "Homo sapiens"}.
-#' @param type shall only the file be retrieved (default) \code{type = "download"} or should the corresponding file be downloaded and subsequently be imported \code{type = "import"}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' file shall be stored. Default is \code{path} = \code{file.path("_ncbi_downloads","genomeassembly_stats")}.
+#' @description  Main genome assembly stats retrieval function for an organism
+#' of interest. By specifying the scientific name of an organism of interest the
+#' corresponding  genome assembly stats file storing the assembly statistics of 
+#' the organism of interest can be downloaded and stored locally. 
+#' Genome assembly stats files can be retrieved from several databases.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"} or \code{db = "genbank"}.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest, e.g. \code{organism = "Homo sapiens"}.
+#' @param type shall only the file be retrieved (default) 
+#' \code{type = "download"} or should the corresponding file be downloaded and 
+#' subsequently be imported \code{type = "import"}.
+#' @param path a character string specifying the location (a folder) in
+#' which the corresponding file shall be stored. Default is 
+#' \code{path} = \code{file.path("_ncbi_downloads","genomeassembly_stats")}.
 #' @author Hajk-Georg Drost
 #' @details Internally this function loads the the overview.txt file from NCBI:
 #' 
@@ -14,33 +21,41 @@
 #' 
 #'  genbank: ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/
 #'  
-#' to retrieve available scientific names of organisms and creates a directory '_ncbi_downloads/genomeassembly_stats' to store
+#' to retrieve available scientific names of organisms and creates a directory
+#' '_ncbi_downloads/genomeassembly_stats' to store
 #' the Genome Assembly Stats of interest as text file for future processing.
 #' In case the corresponding fasta file already exists within the
-#' '_ncbi_downloads/genomeassembly_stats' folder and is accessible within the workspace,
-#' no download process will be performed.
+#' '_ncbi_downloads/genomeassembly_stats' folder and is
+#' accessible within the workspace, no download process will be performed.
 #' 
-#' An example genome assembly stats file can be found here: ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/001/405/GCF_000001405.36_GRCh38.p10/GCF_000001405.36_GRCh38.p10_assembly_stats.txt.
+#' An example genome assembly stats file can be found here:
+#' ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/001/405/
+#' GCF_000001405.36_GRCh38.p10/GCF_000001405.36_GRCh38.p10_assembly_stats.txt.
 #' 
 #' @return File path to downloaded genome assembly stats file.
 #' @examples \dontrun{
-#' # download the genome assembly stats file of Homo sapiens from NCBI RefSeq
[…]
