### 2017-04-06 32a84e6
format to 80 line and update examples
