+++ R/getAttributes.R
@@ -2,13 +2,12 @@
 #' @description This funcion queries the BioMart Interface and returns a table
 #' storing all available attributes for a specific dataset.
 #' 
-#' @param mart a character string specifying the database (mart) for which datasets shall be listed.
-#' @param dataset a character string specifying the dataset for which attributes shall be listed.
+#' @param mart a character string specifying the database (mart) 
+#' for which datasets shall be listed.
+#' @param dataset a character string specifying the dataset for which 
+#' attributes shall be listed.
 #' @author Hajk-Georg Drost
 #' @examples
-#' 
-#' \dontrun{
-#' 
 #' # search for available datasets
 #' getMarts()
 #' 
@@ -17,40 +16,42 @@
 #' head(getDatasets(mart = "ENSEMBL_MART_ENSEMBL"), 10)
 #' 
 #' # choose dataset: "hsapiens_gene_ensembl"
-#' head(getAttributes(mart = "ENSEMBL_MART_ENSEMBL", dataset = "hsapiens_gene_ensembl") , 5)
+#' head(getAttributes(mart = "ENSEMBL_MART_ENSEMBL", 
+#'                    dataset = "hsapiens_gene_ensembl") , 5)
 #' 
-#' }
-#' 
-#' @seealso \code{\link{getMarts}}, \code{\link{getDatasets}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+#' @seealso \code{\link{getMarts}}, \code{\link{getDatasets}},
+#' \code{\link{getFilters}}, \code{\link{organismBM}}, 
+#' \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 #' @export
 getAttributes <- function(mart, dataset){
-        
-        if ((!is.character(mart)) || (!is.character(dataset)))
-                stop("Please use a character string as mart or dataset.", call. = FALSE)
+    
+    if ((!is.character(mart)) || (!is.character(dataset)))
+        stop("Please use a character string as mart or dataset.",
+             call. = FALSE)
     
     if (stringr::str_detect(mart, "ENSEMBL"))
         # connect to BioMart API
         url <-
             paste0(
-                "http://ensembl.org/biomart/martservice?type=attributes&dataset=",
+              "http://ensembl.org/biomart/martservice?type=attributes&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
-            )    
+            )
     if (stringr::str_detect(mart, "plants"))
         # connect to BioMart API
         url <-
             paste0(
-                "http://plants.ensembl.org/biomart/martservice?type=attributes&dataset=",
+       "http://plants.ensembl.org/biomart/martservice?type=attributes&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
-            )    
+            )
     if (stringr::str_detect(mart, "fung"))
         # connect to BioMart API
         url <-
             paste0(
-                "http://fungi.ensembl.org/biomart/martservice?type=attributes&dataset=",
+        "http://fungi.ensembl.org/biomart/martservice?type=attributes&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
@@ -60,26 +61,30 @@ getAttributes <- function(mart, dataset){
         # connect to BioMart API
         url <-
             paste0(
-                "http://protist.ensembl.org/biomart/martservice?type=attributes&dataset=",
+      "http://protist.ensembl.org/biomart/martservice?type=attributes&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
-            )    
+            )
     if (stringr::str_detect(mart, "metazoa"))
         # connect to BioMart API
         url <-
             paste0(
[…]
