+++ R/getCDS.R
@@ -1,11 +1,19 @@
 #' @title Coding Sequence Retrieval
-#' @description Main retrieval function for coding sequences (CDS) of an organism of interest.
-#' By specifying the scientific name of an organism of interest the corresponding fasta-file storing the CDS information for the organism of interest
-#' can be downloaded and stored locally. CDS files can be retrieved from several databases.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"} or \code{db = "ensemblgenomes"}.
-#' @param organism a character string specifying the scientific name of the organism of interest,\code{organism = "Homo sapiens"}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' CDS file shall be stored. Default is \code{path} = \code{file.path("_ncbi_downloads","CDS")}.
+#' @description Main retrieval function for coding sequences (CDS) 
+#' of an organism of interest.
+#' By specifying the scientific name of an organism of interest the 
+#' corresponding fasta-file storing the CDS information for the organism 
+#' of interest can be downloaded and stored locally. CDS files can be retrieved 
+#' from several databases.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: 
+#' \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"} or
+#' \code{db = "ensemblgenomes"}.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest,\code{organism = "Homo sapiens"}.
+#' @param path a character string specifying the location (a folder) 
+#' in which the corresponding CDS file shall be stored. 
+#' Default is \code{path} = \code{file.path("_ncbi_downloads","CDS")}.
 #' @author Hajk-Georg Drost
 #' @return File path to downloaded CDS file.
 #' @examples 
@@ -19,287 +27,392 @@
 #' Ath_CDS <- read_cds(file_path, format = "fasta")
 #' 
 #' }
-#' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
+#' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, 
+#' \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, 
+#' \code{\link{read_cds}}
 #' @export
 
-getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","CDS")){
-    
-    if (!is.element(db, c("refseq", "genbank", "ensembl", "ensemblgenomes")))
-        stop("Please select one of the available data bases: 'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.", call. = FALSE)
-    
-    if (is.element(db, c("refseq", "genbank"))) {
-        # get Kingdom Assembly Summary file
-        AssemblyFilesAllKingdoms <- getKingdomAssemblySummary(db = db)
+getCDS <-
+    function(db = "refseq",
+             organism,
+             path = file.path("_ncbi_downloads", "CDS")) {
+        if (!is.element(db, c("refseq", "genbank", 
+                              "ensembl", "ensemblgenomes")))
+            stop(
+                "Please select one of the available data bases: 
+                'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.",
+                call. = FALSE
+            )
         
-        # test wheter or not genome is available
-        is.genome.available(organism = organism, db = db)
-        
-        if (!file.exists(path)) {
-            dir.create(path, recursive = TRUE)
-        }
-        
[…]
