+++ R/getDatasets.R
@@ -1,71 +1,116 @@
 #' @title Retrieve All Available Datasets for a BioMart Database
 #' @description This funcion queries the BioMart API and returns a table
 #' storing all available datasets for a selected BioMart databases.
-#' 
-#' @param mart a character string specifying the database (mart) for which datasets shall be listed.
+#' @param mart a character string specifying the database (mart) for 
+#' which datasets shall be listed.
 #' @author Hajk-Georg Drost
 #' @examples
-#' \dontrun{
 #' # search for available datasets
 #' getMarts()
-#' 
-#' 
 #' # choose database: "ENSEMBL_MART_ENSEMBL"
 #' head(getDatasets("ENSEMBL_MART_ENSEMBL"), 10)
-#' 
-#' }
-#' 
-#' @seealso \code{\link{getMarts}}, \code{\link{getAttributes}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+#' @seealso \code{\link{getMarts}}, \code{\link{getAttributes}}, 
+#' \code{\link{getFilters}}, \code{\link{organismBM}}, 
+#' \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 #' @export
-getDatasets <- function(mart){
-        
-        if (!is.character(mart))
-                stop("Please use a character string as mart.", call. = FALSE)
-        
-        
+
+getDatasets <- function(mart) {
+    if (!is.character(mart))
+        stop("Please use a character string as mart.", call. = FALSE)
+    
     if (stringr::str_detect(mart, "ENSEMBL"))
         # connect to BioMart API
-        datasetPage <- httr::handle(paste0("http://www.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",mart))
+        datasetPage <-
+            httr::handle(
+                paste0(
+"http://www.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",
+                    mart
+                )
+            )
     
     if (stringr::str_detect(mart, "plants"))
         # connect to BioMart API
-        datasetPage <- httr::handle(paste0("http://plants.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",mart))
+        datasetPage <-
+            httr::handle(
+                paste0(
+"http://plants.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",
+                    mart
+                )
+            )
     
     if (stringr::str_detect(mart, "fung"))
         # connect to BioMart API
-        datasetPage <- httr::handle(paste0("http://fungi.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",mart))
+        datasetPage <-
+            httr::handle(
+                paste0(
+"http://fungi.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",
+                    mart
+                )
+            )
     
     if (stringr::str_detect(mart, "protist"))
         # connect to BioMart API
-        datasetPage <- httr::handle(paste0("http://protists.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",mart))
+        datasetPage <-
+            httr::handle(
+                paste0(
+"http://protists.ensembl.org/biomart/martservice?type=datasets&requestid=biomart&mart=",
+                    mart
+                )
+            )
     
     if (stringr::str_detect(mart, "metazoa"))
         # connect to BioMart API
-        datasetPage <- httr::handle(paste0("http://metazoa.ensembl.org:80/biomart/martservice?type=datasets&requestid=biomart&mart=",mart))
+        datasetPage <-
+            httr::handle(
+                paste0(
[…]
