+++ R/getENSEMBL.Annotation.R
@@ -1,5 +1,6 @@
 #' @title Helper function for retrieving gff files from ENSEMBL
-#' @description This function downloads gff files of query organisms from ENSEMBL.
+#' @description This function downloads gff 
+#' files of query organisms from ENSEMBL.
 #' @param organism scientific name of the organism of interest.
 #' @param type specification type.
 #' @param id.type id type.
@@ -7,161 +8,188 @@
 #' @author Hajk-Georg Drost
 #' @noRd
 
-getENSEMBL.Annotation <- function(organism, type = "dna", id.type = "toplevel", path) {
-    
-    if (!is.element(type, c("dna", "cds", "pep")))
-        stop("Please a 'type' argument supported by this function: 'dna', 'cds', 'pep'.")
-    
-    # test if REST API is responding
-    is.ensembl.alive()
-    
-    new.organism <- stringr::str_replace_all(organism, " ", "_")
-    
-    # if (file.exists(file.path(tempdir(), "ensembl_summary.txt"))) {
-    #     ensembl.available.organisms <-
-    #                          readr::read_delim(
-    #                              file.path(tempdir(), "ensembl_summary.txt"),
-    #                              col_names = c(
-    #                                  "division",
-    #                                  "taxon_id",
-    #                                  "name",
-    #                                  "release",
-    #                                  "display_name",
-    #                                  "accession",
-    #                                  "common_name",
-    #                                  "assembly"
-    #                              ),
-    #                              col_types = readr::cols(
-    #                                  division = readr::col_character(),
-    #                                  taxon_id = readr::col_integer(),
-    #                                  name = readr::col_character(),
-    #                                  release = readr::col_integer(),
-    #                                  display_name = readr::col_character(),
-    #                                  accession = readr::col_character(),
-    #                                  common_name = readr::col_character(),
-    #                                  assembly = readr::col_character()
-    #                              ),
-    #                              comment = "#",
-    #                              delim = "\t"
-    #                          )
-    # }
-    # 
-    if (!file.exists(file.path(tempdir(), "ensembl_summary.txt"))) {
-       # check if organism is available on ENSEMBL
-       tryCatch({
+getENSEMBL.Annotation <-
+    function(organism,
+             type = "dna",
+             id.type = "toplevel",
+             path) {
+        if (!is.element(type, c("dna", "cds", "pep")))
+            stop("Please a 'type' argument supported by this function: 
+                 'dna', 'cds', 'pep'.")
+        
+        # test if REST API is responding
+        is.ensembl.alive()
+        
+        new.organism <- stringr::str_replace_all(organism, " ", "_")
+        
+        # if (file.exists(file.path(tempdir(), "ensembl_summary.txt"))) {
+        #     ensembl.available.organisms <-
+        #                          readr::read_delim(
+        #                           file.path(tempdir(), "ensembl_summary.txt"),
+        #                              col_names = c(
+        #                                  "division",
+        #                                  "taxon_id",
+        #                                  "name",
+        #                                  "release",
+        #                                  "display_name",
+        #                                  "accession",
+        #                                  "common_name",
+        #                                  "assembly"
+        #                              ),
+        #                              col_types = readr::cols(
+        #                                  division = readr::col_character(),
+        #                                  taxon_id = readr::col_integer(),
+        #                                  name = readr::col_character(),
+        #                                  release = readr::col_integer(),
+        #                                 display_name = readr::col_character(),
+        #                                  accession = readr::col_character(),
+        #                                  common_name = readr::col_character(),
+        #                                  assembly = readr::col_character()
+        #                              ),
+        #                              comment = "#",
[…]
