+++ man/getAttributes.Rd
@@ -7,18 +7,17 @@
 getAttributes(mart, dataset)
 }
 \arguments{
-\item{mart}{a character string specifying the database (mart) for which datasets shall be listed.}
+\item{mart}{a character string specifying the database (mart) 
+for which datasets shall be listed.}
 
-\item{dataset}{a character string specifying the dataset for which attributes shall be listed.}
+\item{dataset}{a character string specifying the dataset for which 
+attributes shall be listed.}
 }
 \description{
 This funcion queries the BioMart Interface and returns a table
 storing all available attributes for a specific dataset.
 }
 \examples{
-
-\dontrun{
-
 # search for available datasets
 getMarts()
 
@@ -27,13 +26,14 @@ getMarts()
 head(getDatasets(mart = "ENSEMBL_MART_ENSEMBL"), 10)
 
 # choose dataset: "hsapiens_gene_ensembl"
-head(getAttributes(mart = "ENSEMBL_MART_ENSEMBL", dataset = "hsapiens_gene_ensembl") , 5)
-
-}
+head(getAttributes(mart = "ENSEMBL_MART_ENSEMBL", 
+                   dataset = "hsapiens_gene_ensembl") , 5)
 
 }
 \seealso{
-\code{\link{getMarts}}, \code{\link{getDatasets}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+\code{\link{getMarts}}, \code{\link{getDatasets}},
+\code{\link{getFilters}}, \code{\link{organismBM}}, 
+\code{\link{organismFilters}}, \code{\link{organismAttributes}}
 }
 \author{
 Hajk-Georg Drost
