+++ R/getENSEMBL.Seq.R
@@ -1,5 +1,6 @@
 #' @title Helper function for retrieving biological sequence files from ENSEMBL
-#' @description This function downloads gff files of query organisms from ENSEMBL.
+#' @description This function downloads gff files of query 
+#' organisms from ENSEMBL.
 #' @param organism scientific name of the organism of interest.
 #' @param type biological sequence type.
 #' @param id.type id type.
@@ -10,7 +11,8 @@
 getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
     
     if (!is.element(type, c("dna", "cds", "pep", "ncrna")))
-        stop("Please a 'type' argument supported by this function: 'dna', 'cds', 'pep', 'ncrna'.")
+        stop("Please a 'type' argument supported by this function: 
+             'dna', 'cds', 'pep', 'ncrna'.")
     
     new.organism <- stringr::str_replace_all(organism, " ", "_")
     
@@ -51,10 +53,15 @@ getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
         # check if organism is available on ENSEMBL
         tryCatch({
             ensembl.available.organisms <-
-                jsonlite::fromJSON("http://rest.ensembl.org/info/species?content-type=application/json")
+                jsonlite::fromJSON(
+           "http://rest.ensembl.org/info/species?content-type=application/json")
         }, error = function(e) {
             warning(
-                "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensembl.org' currently available? Do you have a fast and stable internet connection?", call. = FALSE
+                "The API 'http://rest.ensembl.org' does not seem to work 
+                properly. Are you connected to the internet? Is the homepage
+                'http://rest.ensembl.org' currently available? 
+                Do you have a fast and stable internet connection?", 
+                call. = FALSE
             )
             return(FALSE)
         }
@@ -64,7 +71,8 @@ getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
         
         # transform list object returned by 'fromJSON' to tibble
         ensembl.available.organisms <-
-            tibble::as_tibble(dplyr::select(ensembl.available.organisms$species, -aliases, -groups))
+            tibble::as_tibble(dplyr::select(ensembl.available.organisms$species,
+                                            -aliases, -groups))
         
         readr::write_tsv(ensembl.available.organisms,
                          file.path(tempdir(), "ensembl_summary.txt"))
@@ -76,12 +84,13 @@ getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
         warning(
             "Unfortunately organism '",
             organism,
-            "' is not available at ENSEMBL. Please check whether or not the organism name is typed correctly. Thus, download of this species has been omitted."
+            "' is not available at ENSEMBL. Please check whether or not 
+            the organism name is typed correctly. Thus, download of this 
+            species has been omitted."
         )
         return(FALSE)
     }
         
-    
     # test proper API access
     tryCatch({
         json.qry.info <-
@@ -94,7 +103,9 @@ getENSEMBL.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
             )
     }, error = function(e) {
         warning(
-            "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensembl.org' currently available?"
+            "The API 'http://rest.ensembl.org' does not seem to work properly. 
+            Are you connected to the internet? Is the homepage 
+            'http://rest.ensembl.org' currently available?"
         )
         return(FALSE)
[…]
