+++ R/getENSEMBLGENOMES.Seq.R
@@ -7,302 +7,364 @@
 #' @author Hajk-Georg Drost
 #' @noRd
 
-getENSEMBLGENOMES.Seq <- function(organism, type = "dna", id.type = "toplevel", path) {
-    
-    if (!is.element(type, c("dna", "cds", "pep", "ncrna")))
-        stop("Please a 'type' argument supported by this function: 'dna', 'cds', 'pep', 'ncrna'.")
-    
-    new.organism <- stringr::str_replace_all(organism, " ", "_")
-    name <- NULL
-    # test if REST API is responding
-    is.ensemblgenomes.alive()
-    
-    if (file.exists(file.path(tempdir(), "ensemblgenomes_summary.txt"))) {
-        suppressWarnings(
-            ensembl.available.organisms <-
-                readr::read_delim(
-                    file.path(tempdir(), "ensemblgenomes_summary.txt"),
-                    delim = "\t",
-                    quote = "\"",
-                    escape_backslash = FALSE,
-                    col_names = c(
-                        "division",
-                        "taxon_id",
-                        "name",
-                        "release",
-                        "display_name",
-                        "accession",
-                        "common_name",
-                        "assembly"
-                    ),
-                    col_types = readr::cols(
-                        division = readr::col_character(),
-                        taxon_id = readr::col_integer(),
-                        name = readr::col_character(),
-                        release = readr::col_integer(),
-                        display_name = readr::col_character(),
-                        accession = readr::col_character(),
-                        common_name = readr::col_character(),
-                        assembly = readr::col_character()
-                    ),
-                    comment = "#"
+getENSEMBLGENOMES.Seq <-
+    function(organism,
+             type = "dna",
+             id.type = "toplevel",
+             path) {
+        
+        if (!is.element(type, c("dna", "cds", "pep", "ncrna")))
+            stop("Please a 'type' argument supported by this function: 
+                 'dna', 'cds', 'pep', 'ncrna'.")
+        
+        new.organism <- stringr::str_replace_all(organism, " ", "_")
+        name <- NULL
+        # test if REST API is responding
+        is.ensemblgenomes.alive()
+        
+        if (file.exists(file.path(tempdir(), "ensemblgenomes_summary.txt"))) {
+            suppressWarnings(
+                ensembl.available.organisms <-
+                    readr::read_delim(
+                        file.path(tempdir(), "ensemblgenomes_summary.txt"),
+                        delim = "\t",
+                        quote = "\"",
+                        escape_backslash = FALSE,
+                        col_names = c(
+                            "division",
+                            "taxon_id",
+                            "name",
+                            "release",
+                            "display_name",
+                            "accession",
+                            "common_name",
+                            "assembly"
+                        ),
+                        col_types = readr::cols(
+                            division = readr::col_character(),
+                            taxon_id = readr::col_integer(),
+                            name = readr::col_character(),
+                            release = readr::col_integer(),
+                            display_name = readr::col_character(),
+                            accession = readr::col_character(),
+                            common_name = readr::col_character(),
+                            assembly = readr::col_character()
+                        ),
+                        comment = "#"
+                    )
+            )
+        }
+        
+        if (!file.exists(file.path(tempdir(), "ensemblgenomes_summary.txt"))) {
+            # check if organism is available on ENSEMBL
+            tryCatch({
+                ensembl.available.organisms <-
+                    jsonlite::fromJSON(
+     "http://rest.ensemblgenomes.org/info/species?content-type=application/json"
+                    )
+            }, error = function(e) {
+                warning(
+                    "The API 'http://rest.ensemblgenomes.org' does not seem to 
+                    work properly. Are you connected to the internet? 
+                    Is the homepage 'http://rest.ensemblgenomes.org' 
+                    currently available?",
+                    call. = FALSE
                 )
-        )
-    }
-    
[…]
