+++ R/getFilters.R
@@ -1,31 +1,29 @@
 #' @title Retrieve All Available Filters for a Specific Dataset
 #' @description This funcion queries the BioMart API and returns a table
 #' storing all available filters for a specific dataset.
-#' 
-#' @param mart a character string specifying the database (mart) for which datasets shall be listed.
-#' @param dataset a character string specifying the dataset for which filters shall be listed.
+#' @param mart a character string specifying the database (mart) for which 
+#' datasets shall be listed.
+#' @param dataset a character string specifying the dataset for which filters 
+#' shall be listed.
 #' @author Hajk-Georg Drost
 #' @examples
-#' \dontrun{
 #' # search for available datasets
 #' getMarts()
-#' 
-#' 
 #' # choose database (mart): "ENSEMBL_MART_ENSEMBL"
 #' head(getDatasets(mart = "ENSEMBL_MART_ENSEMBL"), 10)
-#' 
 #' # choose dataset: "hsapiens_gene_ensembl"
-#' head(getFilters(mart = "ENSEMBL_MART_ENSEMBL", dataset = "hsapiens_gene_ensembl") , 5)
-#' 
-#' }
-#' 
-#' @seealso \code{\link{getMarts}}, \code{\link{getDatasets}}, \code{\link{getAttributes}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+#' head(getFilters(mart = "ENSEMBL_MART_ENSEMBL", 
+#'                 dataset = "hsapiens_gene_ensembl") , 5)
+#' @seealso \code{\link{getMarts}}, \code{\link{getDatasets}}, 
+#' \code{\link{getAttributes}}, \code{\link{organismBM}}, 
+#' \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 #' @export
+
 getFilters <- function(mart, dataset){
-        
-        if ((!is.character(mart)) || (!is.character(dataset)))
-                stop("Please use a character string as mart or dataset.")
-        
+    
+    if ((!is.character(mart)) || (!is.character(dataset)))
+        stop("Please use a character string as mart or dataset.")
+    
     if (stringr::str_detect(mart, "ENSEMBL"))
         # connect to BioMart API
         url <-
@@ -34,21 +32,21 @@ getFilters <- function(mart, dataset){
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
-            )    
+            )
     if (stringr::str_detect(mart, "plants"))
         # connect to BioMart API
         url <-
             paste0(
-                "http://plants.ensembl.org/biomart/martservice?type=filters&dataset=",
+          "http://plants.ensembl.org/biomart/martservice?type=filters&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
-            )    
+            )
     if (stringr::str_detect(mart, "fung"))
         # connect to BioMart API
         url <-
             paste0(
-                "http://fungi.ensembl.org/biomart/martservice?type=filters&dataset=",
+           "http://fungi.ensembl.org/biomart/martservice?type=filters&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
@@ -58,26 +56,32 @@ getFilters <- function(mart, dataset){
         # connect to BioMart API
         url <-
             paste0(
-                "http://protist.ensembl.org/biomart/martservice?type=filters&dataset=",
+         "http://protist.ensembl.org/biomart/martservice?type=filters&dataset=",
                 dataset,
                 "&requestid=biomart&mart=",
                 mart
-            )    
+            )
     if (stringr::str_detect(mart, "metazoa"))
         # connect to BioMart API
         url <-
             paste0(
[…]
