+++ R/getGFF.R
@@ -1,11 +1,17 @@
 #' @title Genome Annotation Retrieval
-#' @description  Main retrieval function for GFF files of an organism of interest.
-#' By specifying the scientific name of an organism of interest the corresponding gff file storing the annotation  for the organism of interest
-#' can be downloaded and stored locally. GFF files can be retrieved from several databases.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, or \code{db = "ensembl"}.
-#' @param organism a character string specifying the scientific name of the organism of interest, e.g. \code{organism = "Homo sapiens"}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' annotation file shall be stored. Default is \code{path = file.path("_ncbi_downloads","genomes")}.
+#' @description  Main retrieval function for GFF files of an 
+#' organism of interest. By specifying the scientific name of an organism of 
+#' interest the corresponding gff file storing the annotation  for the organism 
+#' of interest can be downloaded and stored locally. GFF files can be retrieved 
+#' from several databases.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
+#' or \code{db = "ensembl"}.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest, e.g. \code{organism = "Homo sapiens"}.
+#' @param path a character string specifying the location (a folder) in which 
+#' the corresponding annotation file shall be stored. Default is 
+#' \code{path = file.path("_ncbi_downloads","genomes")}.
 #' @author Hajk-Georg Drost
 #' @details Internally this function loads the the overview.txt file from NCBI:
 #' 
@@ -20,7 +26,6 @@
 #' no download process will be performed.
 #' @return File path to downloaded annotation file.
 #' @examples \dontrun{
-#' 
 #' # download the annotation of Arabidopsis thaliana from refseq
 #' # and store the corresponding genome file in '_ncbi_downloads/annotation'
 #' getGFF( db       = "refseq", 
@@ -36,20 +41,24 @@
 #' 
 #' }
 #' 
-#' @seealso \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGenome}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}
+#' @seealso \code{\link{getProteome}}, \code{\link{getCDS}}, 
+#' \code{\link{getGenome}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}
 #' @export
 
 getGFF <-
     function(db = "refseq",
              organism,
              path = file.path("_ncbi_downloads", "annotation")) {
-        
-        if (!is.element(db, c("refseq", "genbank","ensembl","ensemblgenomes")))
-            stop("Please select one of the available data bases: 'refseq', 'genbank', 'ensembl', 'ensemblgenomes'.")
+       if (!is.element(db, c("refseq", "genbank", "ensembl", "ensemblgenomes")))
+            stop(
+                "Please select one of the available data bases: 'refseq',
+                'genbank', 'ensembl', 'ensemblgenomes'."
+            )
         
         if (is.element(db, c("refseq", "genbank"))) {
             # get Kingdom Assembly Summary file
-            AssemblyFilesAllKingdoms <- getKingdomAssemblySummary(db = db)
+            AssemblyFilesAllKingdoms <-
+                getKingdomAssemblySummary(db = db)
             
             # test wheter or not genome is available
             is.genome.available(organism = organism, db = db)
@@ -61,8 +70,10 @@ getGFF <-
             organism_name <-
                 refseq_category <- version_status <- NULL
             
-            organism <- stringr::str_replace_all(organism,"\\(","")
[…]
