+++ R/getGenome.R
@@ -1,11 +1,17 @@
 #' @title Genome Retrieval
 #' @description  Main genome retrieval function for an organism of interest.
-#' By specifying the scientific name of an organism of interest the corresponding fasta-file storing the genome of the organism of interest
-#' can be downloaded and stored locally. Genome files can be retrieved from several databases.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"}, or \code{db = "ensemblgenomes"}.
-#' @param organism a character string specifying the scientific name of the organism of interest, e.g. \code{organism = "Homo sapiens"}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' genome shall be stored. Default is \code{path} = \code{file.path("_ncbi_downloads","genomes")}.
+#' By specifying the scientific name of an organism of interest the 
+#' corresponding fasta-file storing the genome of the organism of interest
+#' can be downloaded and stored locally. Genome files can be retrieved from 
+#' several databases.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
+#' \code{db = "ensembl"}, or \code{db = "ensemblgenomes"}.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest, e.g. \code{organism = "Homo sapiens"}.
+#' @param path a character string specifying the location (a folder) in which 
+#' the corresponding genome shall be stored. Default is 
+#' \code{path} = \code{file.path("_ncbi_downloads","genomes")}.
 #' @author Hajk-Georg Drost
 #' @details Internally this function loads the the overview.txt file from NCBI:
 #' 
@@ -39,21 +45,27 @@
 #' Ath_genome <- read_genome(file_path, format = "fasta")
 #' }
 #' 
-#' @seealso \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_genome}}
+#' @seealso \code{\link{getProteome}}, \code{\link{getCDS}}, 
+#' \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, 
+#' \code{\link{read_genome}}
 #' @export
 
 getGenome <-
     function(db = "refseq",
              organism,
              path = file.path("_ncbi_downloads", "genomes")) {
-        
-        if (!is.element(db, c("refseq", "genbank","ensembl", "ensemblgenomes")))
-            stop("Please select one of the available data bases: 'refseq', 'genbank', 'ensembl', or 'ensemblgenomes'.", call. = FALSE)
+       if (!is.element(db, c("refseq", "genbank", "ensembl", "ensemblgenomes")))
+            stop(
+                "Please select one of the available data bases: 'refseq',
+                'genbank', 'ensembl', or 'ensemblgenomes'.",
+                call. = FALSE
+            )
         
         if (is.element(db, c("refseq", "genbank"))) {
             # get Kingdom Assembly Summary file
-            AssemblyFilesAllKingdoms <- getKingdomAssemblySummary(db = db)
-        
+            AssemblyFilesAllKingdoms <-
+                getKingdomAssemblySummary(db = db)
+            
             # test wheter or not genome is available
             is.genome.available(organism = organism, db = db)
             
@@ -64,8 +76,10 @@ getGenome <-
             organism_name <-
                 refseq_category <- version_status <- NULL
             
-            organism <- stringr::str_replace_all(organism,"\\(","")
-            organism <- stringr::str_replace_all(organism,"\\)","")
+            organism <-
+                stringr::str_replace_all(organism, "\\(", "")
+            organism <-
[…]
