+++ R/getKingdoms.R
@@ -1,23 +1,25 @@
 #' @title Retrieve available kingdoms of life
 #' @description A short list of available kingdoms of life
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"}, \code{db = "ensemblgenomes"}.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
+#' \code{db = "ensembl"}, \code{db = "ensemblgenomes"}. 
 #' Default is \code{db = "refseq"}.
 #' @author Hajk-Georg Drost
 #' @examples
-#' \dontrun{ 
 #' # retrieve kingdoms available from refseq
 #' getKingdoms(db = "refseq")
 #' 
 #' # retrieve kingdoms available from genbank
 #' getKingdoms(db = "genbank")
-#' }
-#' @seealso \code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGroups}}
+#' @seealso \code{\link{meta.retrieval}}, \code{\link{getGenome}}, 
+#' \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getGroups}}
 #' @export
 
 getKingdoms <- function(db = "refseq"){
     
     if (!is.element(db, c("refseq", "genbank","ensembl", "ensemblgenomes")))
-        stop("Please select one of the available data bases: 'refseq', 'genbank', 'ensembl', or 'ensemblgenomes'.", call. = FALSE)
+        stop("Please select one of the available data bases: 'refseq', 
+             'genbank', 'ensembl', or 'ensemblgenomes'.", call. = FALSE)
     
     if (db == "refseq") {
         return(
