+++ R/getMarts.R
@@ -1,16 +1,13 @@
-#' @title Retrieve All Available BioMart Databases
-#' @description This funcion queries the BioMart API and returns a table
-#' storing all available BioMart databases.
-#' 
+#' @title Retrieve information about available Ensembl Biomart databases 
+#' @description This funcion queries the Ensembl Biomart API and returns a table
+#' storing information about all available Ensembl Biomart databases.
 #' @author Hajk-Georg Drost
 #' @examples
-#' 
-#' \dontrun{
-#' # get a table of all available databases from BioMart
+#' # get a table of all available databases from Ensembl Biomart
 #'  getMarts()
-#' }
-#' 
-#' @seealso \code{\link{getDatasets}}, \code{\link{getAttributes}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+#' @seealso \code{\link{getDatasets}}, \code{\link{getAttributes}}, 
+#' \code{\link{getFilters}}, \code{\link{organismBM}}, 
+#' \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 #' @export
 getMarts <- function(){
     
