+++ R/getMetaGenomeAnnotations.R
@@ -1,97 +1,113 @@
 #' @title Retrieve annotation *.gff files for metagenomes from NCBI Genbank
-#' @description Retrieve available annotation *.gff files for metagenomes from NCBI Genbank. NCBI Genbank allows users
-#' to download entire metagenomes and their annotations of several metagenome projects. This function downloads
-#'  available metagenomes that can then be downloaded via \code{\link{getMetaGenomes}}.
+#' @description Retrieve available annotation *.gff files for metagenomes 
+#' from NCBI Genbank. NCBI Genbank allows users
+#' to download entire metagenomes and their annotations of several metagenome 
+#' projects. This function downloads available metagenomes that can then be 
+#' downloaded via \code{\link{getMetaGenomes}}.
 #' @param name metagenome name retrieved by \code{\link{listMetaGenomes}}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' metagenome annotations shall be stored. Default is \code{path} = \code{file.path("_ncbi_downloads","metagenome","annotations")}.
+#' @param path a character string specifying the location (a folder) 
+#' in which the corresponding metagenome annotations shall be stored. 
+#' Default is
+#' \code{path} = \code{file.path("_ncbi_downloads","metagenome","annotations")}.
 #' @author Hajk-Georg Drost
 #' @examples
-#' \dontrun{
 #' # Frist, retrieve a list of available metagenomes
 #' listMetaGenomes()
-#' 
+#' \dontrun{
 #' # Now, retrieve the 'human gut metagenome'
 #' getMetaGenomeAnnotations(name = "human gut metagenome")
 #' }
-#' @seealso \code{\link{getMetaGenomes}}, \code{\link{listMetaGenomes}}, \code{\link{getGFF}}  
+#' @seealso \code{\link{getMetaGenomes}}, \code{\link{listMetaGenomes}}, 
+#' \code{\link{getGFF}}  
 #' @export
 
-getMetaGenomeAnnotations <- function(name, path = file.path("_ncbi_downloads","metagenome","annotations")) {
-    
-    if (!is.element(name, listMetaGenomes(details = FALSE)))
-        stop("Unfortunately, the metagenome '",name,"' is not available. Please consult the listMetaGenomes() function for available metagenomes.")
-    
-    if (!file.exists(path)) {
-        dir.create(path, recursive = TRUE)
-    }
-    
-    organism_name <- NULL
-    
-    # retrieve metagenomes assembly_summary.txt file
-    mgs <- getMetaGenomeSummary()
-    
-    metagenomes.members <- dplyr::filter(mgs, organism_name == name)
-    file.names <- metagenomes.members$ftp_path
-    
-    for (i in seq_len(length(file.names))) {
-        
-        download_url <-
-            paste0(
-                file.names[i],
-                "/",
-                paste0(
-                    metagenomes.members$assembly_accession[i],
-                    "_",
-                    metagenomes.members$asm_name[i],
-                    "_genomic.gff.gz"
-                )
+getMetaGenomeAnnotations <-
+    function(name,
+             path = file.path("_ncbi_downloads", "metagenome", "annotations")) {
+        if (!is.element(name, listMetaGenomes(details = FALSE)))
+            stop(
+                "Unfortunately, the metagenome '",
+                name,
+                "' is not available. Please consult the listMetaGenomes() 
+                function for available metagenomes."
             )
         
-        tryCatch({utils::capture.output(downloader::download(
-            download_url,
-            destfile = file.path(path, paste0(basename(file.names[i]), "_genomic.gff.gz")),
-            mode = "wb"
-        ))}, error = function(e)
-            stop(
[…]
