+++ R/getProteome.R
@@ -1,11 +1,17 @@
 #' @title Proteome Retrieval
 #' @description Main proteome retrieval function for an organism of interest.
-#' By specifying the scientific name of an organism of interest the corresponding fasta-file storing the proteome of the organism of interest
-#' can be downloaded and stored locally. Proteome files can be retrieved from several databases.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"} or \code{db = "ensemblgenomes"}.
-#' @param organism a character string specifying the scientific name of the organism of interest, e.g. \code{organism = "Homo sapiens"}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' proteome shall be stored. Default is \code{path} = \code{file.path("_ncbi_downloads","proteomes")}.
+#' By specifying the scientific name of an organism of interest the 
+#' corresponding fasta-file storing the proteome of the organism of interest
+#' can be downloaded and stored locally. Proteome files can be retrieved from 
+#' several databases.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
+#' \code{db = "ensembl"} or \code{db = "ensemblgenomes"}.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest, e.g. \code{organism = "Homo sapiens"}.
+#' @param path a character string specifying the location (a folder) in which 
+#' the corresponding proteome shall be stored. Default is 
+#' \code{path} = \code{file.path("_ncbi_downloads","proteomes")}.
 #' @author Hajk-Georg Drost
 #' @details Internally this function loads the the overview.txt file from NCBI:
 #' 
@@ -34,118 +40,257 @@
 #' 
 #' Ath_proteome <- read_proteome(file_path, format = "fasta")
 #' }
-#' @seealso \code{\link{getGenome}}, \code{\link{getCDS}}, \code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, \code{\link{read_proteome}}
+#' @seealso \code{\link{getGenome}}, \code{\link{getCDS}}, \code{\link{getGFF}},
+#' \code{\link{getRNA}}, \code{\link{meta.retrieval}}, 
+#' \code{\link{read_proteome}}
 #' @export
 
-getProteome <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","proteomes")){
-    
-    if (!is.element(db, c("refseq", "genbank", "ensembl", "ensemblgenomes")))
-        stop("Please select one of the available data bases: 'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.", call. = FALSE)
-    
-    if (is.element(db, c("refseq", "genbank"))) {
-        # get Kingdom Assembly Summary file
-        AssemblyFilesAllKingdoms <- getKingdomAssemblySummary(db = db)
-        
-        # test wheter or not genome is available
-        is.genome.available(organism = organism, db = db)
-        
-        if (!file.exists(path)) {
-            dir.create(path, recursive = TRUE)
-        }
-        
-        organism_name <- refseq_category <- version_status <- NULL
-        organism <- stringr::str_replace_all(organism,"\\(","")
-        organism <- stringr::str_replace_all(organism,"\\)","")
-        
-        FoundOrganism <-
-            dplyr::filter(
-                AssemblyFilesAllKingdoms,
-                stringr::str_detect(organism_name, organism),
-                ((refseq_category == "representative genome") ||
-                     (refseq_category == "reference genome")
-                ),
-                (version_status == "latest")
+getProteome <-
+    function(db = "refseq",
+             organism,
+             path = file.path("_ncbi_downloads", "proteomes")) {
[…]
