+++ man/getENSEMBLInfo.Rd
@@ -7,7 +7,8 @@
 getENSEMBLInfo()
 }
 \description{
-Retrieve species and genome information from http://rest.ensembl.org/info/species?content-type=application/json/.
+Retrieve species and genome information from 
+http://rest.ensembl.org/info/species?content-type=application/json/.
 }
 \author{
 Hajk-Georg Drost
