+++ man/getGENOMEREPORT.Rd
@@ -7,7 +7,8 @@
 getGENOMEREPORT()
 }
 \description{
-Retrieves NCBI GENOME_REPORTS file from ftp://ftp.ncbi.nlm.nih.gov/genomes/GENOME_REPORTS/overview.txt.
+Retrieves NCBI GENOME_REPORTS file from 
+ftp://ftp.ncbi.nlm.nih.gov/genomes/GENOME_REPORTS/overview.txt.
 }
 \author{
 Hajk-Georg Drost
