+++ man/getKingdomAssemblySummary.Rd
@@ -2,7 +2,8 @@
 % Please edit documentation in R/getKingdomAssemblySummary.R
 \name{getKingdomAssemblySummary}
 \alias{getKingdomAssemblySummary}
-\title{Retrieve and summarise the assembly_summary.txt files from NCBI for all kingdoms}
+\title{Retrieve and summarise the assembly_summary.txt files from 
+NCBI for all kingdoms}
 \usage{
 getKingdomAssemblySummary(db)
 }
@@ -10,7 +11,8 @@ getKingdomAssemblySummary(db)
 \item{db}{database name. E.g. \code{refseq} or \code{genbank}.}
 }
 \description{
-Retrieval function of the assembly_summary.txt file from NCBI for all kingdoms.
+Retrieval function of the assembly_summary.txt file 
+from NCBI for all kingdoms.
 The assembly_summary.txt files store available species on NCBI.
 }
 \examples{
