+++ man/getMarts.Rd
@@ -2,24 +2,22 @@
 % Please edit documentation in R/getMarts.R
 \name{getMarts}
 \alias{getMarts}
-\title{Retrieve All Available BioMart Databases}
+\title{Retrieve information about available Ensembl Biomart databases}
 \usage{
 getMarts()
 }
 \description{
-This funcion queries the BioMart API and returns a table
-storing all available BioMart databases.
+This funcion queries the Ensembl Biomart API and returns a table
+storing information about all available Ensembl Biomart databases.
 }
 \examples{
-
-\dontrun{
-# get a table of all available databases from BioMart
+# get a table of all available databases from Ensembl Biomart
  getMarts()
 }
-
-}
 \seealso{
-\code{\link{getDatasets}}, \code{\link{getAttributes}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+\code{\link{getDatasets}}, \code{\link{getAttributes}}, 
+\code{\link{getFilters}}, \code{\link{organismBM}}, 
+\code{\link{organismFilters}}, \code{\link{organismAttributes}}
 }
 \author{
 Hajk-Georg Drost
