+++ R/getRNA.R
@@ -1,11 +1,18 @@
 #' @title RNA Sequence Retrieval
-#' @description Main retrieval function for RNA sequences of an organism of interest.
-#' By specifying the scientific name of an organism of interest the corresponding fasta-file storing the RNA information for the organism of interest
-#' can be downloaded and stored locally. RNA files can be retrieved from several databases.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"} or \code{db = "ensemblgenomes"}.
-#' @param organism a character string specifying the scientific name of the organism of interest,\code{organism = "Homo sapiens"}.
-#' @param path a character string specifying the location (a folder) in which the corresponding
-#' CDS file shall be stored. Default is \code{path} = \code{file.path("_ncbi_downloads","RNA")}.
+#' @description Main retrieval function for RNA sequences of an organism 
+#' of interest. By specifying the scientific name of an organism of interest the
+#' corresponding fasta-file storing the RNA information for the organism 
+#' of interest can be downloaded and stored locally. 
+#' RNA files can be retrieved from several databases.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
+#' \code{db = "ensembl"} or \code{db = "ensemblgenomes"}.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest,\code{organism = "Homo sapiens"}.
+#' @param path a character string specifying the location (a folder) in which 
+#' the corresponding
+#' CDS file shall be stored. Default is 
+#' \code{path} = \code{file.path("_ncbi_downloads","RNA")}.
 #' @author Hajk-Georg Drost
 #' @return File path to downloaded RNA file.
 #' @examples 
@@ -17,297 +24,402 @@
 #'              path     = file.path("_ncbi_downloads","RNA"))
 #' 
 #' Ath_RNA <- read_rna(file_path, format = "fasta")
-#' 
 #' }
-#' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getRNA}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, \code{\link{read_cds}}
+#' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, 
+#' \code{\link{getRNA}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, 
+#' \code{\link{read_cds}}
 #' @export
 
-getRNA <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","RNA")){
-    
-    if (!is.element(db, c("refseq", "genbank", "ensembl", "ensemblgenomes")))
-        stop("Please select one of the available data bases: 'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.", call. = FALSE)
-    
-    if (is.element(db, c("refseq", "genbank"))) {
-        # get Kingdom Assembly Summary file
-        AssemblyFilesAllKingdoms <- getKingdomAssemblySummary(db = db)
+getRNA <-
+    function(db = "refseq",
+             organism,
+             path = file.path("_ncbi_downloads", "RNA")) {
+        if (!is.element(db, c("refseq", "genbank", "ensembl", 
+                              "ensemblgenomes")))
+            stop(
+                "Please select one of the available data bases: 
+                'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.",
+                call. = FALSE
+            )
         
-        # test wheter or not genome is available
-        is.genome.available(organism = organism, db = db)
-        
-        if (!file.exists(path)) {
-            dir.create(path, recursive = TRUE)
-        }
-        
-        organism_name <- refseq_category <- version_status <- NULL
[…]
