+++ R/is.genome.available.R
@@ -1,22 +1,24 @@
 #' @title Check Genome Availability
-#' @description This function checks the availability of a given genome on the NBCI servers specified
-#' as scientific name.
-#' @param organism a character string specifying the scientific name of the organism of interest, e.g. \code{organism = "Homo sapiens"}.
-#' @param details a logical value specifying whether or not details on genome size, kingdom, etc. shall be printed to the
-#' console intead of a boolean value.
-#' @param db a character string specifying the database for which genome availability shall be checked,
-#' e.g. \code{db = "refseq"},\code{db = "genbank"}, \code{db = "ensembl"}, \code{db = "ensemblgenomes"}, or \code{db =  "all"}.
+#' @description This function checks the availability of a given genome on the 
+#' NBCI servers specified as scientific name.
+#' @param organism a character string specifying the scientific name of the 
+#' organism of interest, e.g. \code{organism = "Homo sapiens"}.
+#' @param details a logical value specifying whether or not details on genome 
+#' size, kingdom, etc. shall be printed to the console intead of a 
+#' boolean value.
+#' @param db a character string specifying the database for which genome 
+#' availability shall be checked,
+#' e.g. \code{db = "refseq"},\code{db = "genbank"}, \code{db = "ensembl"}, 
+#' \code{db = "ensemblgenomes"}, or \code{db =  "all"}.
 #' @details
-#'
-#' Internally this function calls the \code{\link{listGenomes}} function to detect all available genomes
-#' and checks whether or not the specified organism is available for download.
-#'
-#' @return a logical value specifing whether or not the genome of the input organism
-#' is available. In case \code{details} = \code{TRUE} only a character string specifying the
-#' genome details is being returned.
+#' Internally this function calls the \code{\link{listGenomes}} function to 
+#' detect all available genomes and checks whether or not the specified organism
+#' is available for download.
+#' @return a logical value specifing whether or not the genome of the input 
+#' organism is available. In case \code{details} = \code{TRUE} only a character 
+#' string specifying the genome details is being returned.
 #' @author Hajk-Georg Drost
 #' @examples \dontrun{
-#'
 #' # checking whether the Homo sapiens genome is stored on NCBI
 #' is.genome.available(organism = "Homo sapiens", db = "refseq")
 #'
@@ -27,262 +29,337 @@
 #' is.genome.available(organism = "Homo sapiens", db = "ensembl")
 #' 
 #' # and printing details
-#' is.genome.available(organism = "Homo sapiens", details = TRUE, db = "ensembl")
+#' is.genome.available(organism = "Homo sapiens",
+#'                     details = TRUE, 
+#'                     db = "ensembl")
 #' }
 #' @export
 
-is.genome.available <- function(organism, details = FALSE, db = "refseq"){
-    
-    if (!is.element(db, c("refseq", "genbank", "ensembl", "ensemblgenomes")))
-        stop("Please select one of the available data bases: 'refseq', 'genbank', or 'ensembl'",
-             call. = FALSE)
-    
-    if (is.element(db, c("refseq", "genbank"))) {
-        
-        aliases <- groups <- NULL
-        # if AssemblyFilesAllKingdoms.txt file was already generated/downloaded then use the local version
-        # stored in temp()
-        if (file.exists(file.path(tempdir(), paste0("AssemblyFilesAllKingdoms_", db, ".txt")))) {
-            suppressWarnings(
-                AssemblyFilesAllKingdoms <-
-                    readr::read_tsv(
-                        file.path(tempdir(), paste0("AssemblyFilesAllKingdoms_", db, ".txt")),
-                        col_names = c(
-                            "assembly_accession",
-                            "bioproject",
-                            "biosample",
-                            "wgs_master",
-                            "refseq_category",
[…]
