+++ R/meta.retrieval.R
@@ -1,46 +1,69 @@
 #' @title Perform Meta-Genome Retieval
-#' @description Download genomes, proteomes, cds, gff, rna, or assembly stats files of all species within a kingdom of life.
-#' @param kingdom a character string specifying the kingdom of the organisms of interest,
-#' e.g. "archaea","bacteria", "fungi", "invertebrate", "plant", "protozoa", "vertebrate_mammalian", or "vertebrate_other".
+#' @description Download genomes, proteomes, cds, gff, rna, or assembly stats 
+#' files of all species within a kingdom of life.
+#' @param kingdom a character string specifying the kingdom of the organisms 
+#' of interest, e.g. "archaea","bacteria", "fungi", "invertebrate", "plant", 
+#' "protozoa", "vertebrate_mammalian", or "vertebrate_other".
 #' Available kingdoms can be retrieved with \code{\link{getKingdoms}}.
-#' @param group only species belonging to this subgroup will be downloaded. Groups can be retrieved with \code{\link{getGroups}}.
-#' @param db a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "emsembl"} or \code{db = "ensemblgenomes"}.
+#' @param group only species belonging to this subgroup will be downloaded. 
+#' Groups can be retrieved with \code{\link{getGroups}}.
+#' @param db a character string specifying the database from which the genome 
+#' shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
+#' \code{db = "emsembl"} or \code{db = "ensemblgenomes"}.
 #' @param type type of sequences that shall be retrieved. Options are:
 #' \itemize{
-#'  \item \code{type = "genome"} (for genome assembly retrieval; see also \code{\link{getGenome}}), 
-#'  \item \code{type = "proteome"} (for proteome retrieval; see also \code{\link{getProteome}}),
-#'  \item \code{type = "CDS"} (for coding sequence retrieval; see also \code{\link{getCDS}}),
-#'  \item \code{type = "gff"} (for annotation file retrieval in gff format; see also \code{\link{getGFF}}),
-#'  \item \code{type = "rna"} (for RNA file retrieval in fasta format; see also \code{\link{getRNA}}),
-#'  \item \code{type = "assemblystats"} (for genome assembly quality stats file retrieval; see also \code{\link{getAssemblyStats}}).
+#'  \item \code{type = "genome"} :
+#'  (for genome assembly retrieval; see also \code{\link{getGenome}}), 
+#'  \item \code{type = "proteome"} :
+#'  (for proteome retrieval; see also \code{\link{getProteome}}),
+#'  \item \code{type = "CDS"} :
+#'  (for coding sequence retrieval; see also \code{\link{getCDS}}),
+#'  \item \code{type = "gff"} :
+#' (for annotation file retrieval in gff format; see also \code{\link{getGFF}}),
+#'  \item \code{type = "rna"} :
+#'  (for RNA file retrieval in fasta format; see also \code{\link{getRNA}}),
+#'  \item \code{type = "assemblystats"} :
+#'  (for genome assembly quality stats file retrieval; 
+#'  see also \code{\link{getAssemblyStats}}).
 #'  }
-#' @param combine just in case \code{type = "assemblystats"} is specified, shall assemby stats of individual species be imported and combined to a \code{\link{data.frame}}? 
-#' @param path path to the folder in which downloaded genomes shall be stored. By default the
-#' kingdom name is used to name the output folder.
+#' @param combine just in case \code{type = "assemblystats"} is specified, shall
+#' assemby stats of individual species be imported and combined to a 
+#' \code{\link{data.frame}}? 
[…]
