+++ R/organismBM.R
@@ -1,35 +1,33 @@
-#' @title Retrieve Biomart Marts and Datasets for an Organism
-#' @description This function returns either all available biomart connections for
-#' all available organisms for which biomart access is possible, or (when specified) returns all
-#' organism specific biomart connections.
-#' @param organism a character string specifying the scientific name of a query organism.
-#' Default is \code{organism} = \code{NULL}. In this case all available biomart connections are returned.
-#' @param update a logical value specifying whether or not the local listMart.txt and listDatasets.txt files shall be updated
-#' by remote access to BioMart. 
+#' @title Retrieve Ensembl Biomart marts and datasets for a query organism
+#' @description This function returns either all available biomart connections 
+#' for all available organisms for which biomart access is possible, or 
+#' (when specified) returns all organism specific biomart connections.
+#' @param organism a character string specifying the scientific name of a 
+#' query organism. Default is \code{organism} = \code{NULL}. In this case all 
+#' available biomart connections are returned.
+#' @param update a logical value specifying whether or not the local 
+#' listMart.txt and listDatasets.txt files shall be updated by remote access
+#'  to BioMart. 
 #' @author Hajk-Georg Drost
 #' @details
-#' This function collects all available biomart connections and returns a table storing
-#' the organism for which biomart connections are available as well as the corresponding mart and database.
-#'   
-#' 
+#' This function collects all available biomart connections and returns a table 
+#' storing the organism for which biomart connections are available as well as 
+#' the corresponding mart and database.
 #' @note
-#' When you run this function for the first time, the data retrieval procedure will take some time,
-#' due to the remote access to BioMart. The corresponding result is then saved in a *.txt file named
-#' "_biomart/listDatasets.txt" in the \code{\link{tempdir}} directory, allowing subsequent queries to perform much faster.
-#' 
-#' @examples \dontrun{
-#' 
+#' When you run this function for the first time, the data retrieval procedure 
+#' will take some time, due to the remote access to BioMart. The corresponding 
+#' result is then saved in a *.txt file named "_biomart/listDatasets.txt" in the
+#' \code{\link{tempdir}} directory, allowing subsequent queries to perform
+#'  much faster.
+#' @examples 
 #' # returning all available biomart connections
 #' head(organismBM(), 20)
-#' 
 #' # retrieving all available datasets and biomart connections for
 #' # a specific query organism (scientific name)
 #' organismBM(organism = "Homo sapiens")
-#' 
-#' # you can also update the downloaded version using the "update = TRUE" argument
+#' # you can also update the downloaded version using 
+#' # the "update = TRUE" argument
 #' head(organismBM(update = TRUE), 20)
-#' 
-#' }
 #' @references
 #' 
 #' \url{http://biomart.org/}
@@ -42,118 +40,131 @@
 #' microarray data analysis. Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean
 #' Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21,
 #' 3439-3440 (2005).
-#' @seealso \code{\link{getMarts}}, \code{\link{getDatasets}}, \code{\link{biomart}},
-#' \code{\link{organismFilters}}, \code{\link{organismAttributes}}
+#' @seealso \code{\link{getMarts}}, \code{\link{getDatasets}}, 
+#' \code{\link{biomart}}, \code{\link{organismFilters}}, 
+#' \code{\link{organismAttributes}}
 #' @export
 
-organismBM <- function(organism = NULL, update = FALSE){
-
-        organism_name <- description <- mart <- dataset <- NULL
+organismBM <- function(organism = NULL, update = FALSE) {
[…]
