+++ R/read_assemblystats.R
@@ -1,16 +1,28 @@
 #' @title Import Genome Assembly Stats File
-#' @description This function reads an organism specific Genome Assembly Stats file that was retrieved with \code{\link{getAssemblyStats}}.
-#' @param file a character string specifying the path to the file storing the Genome Assembly Stats file.
-#' @param type either \code{type = "raw"} to import the entire genome assembly stats file or \code{type = "stats"} to import overall statistics including all chromosomes, mitochondria and plastids.
+#' @description This function reads an organism specific Genome Assembly 
+#' Stats file that was retrieved with \code{\link{getAssemblyStats}}.
+#' @param file a character string specifying the path to the file storing 
+#' the Genome Assembly Stats file.
+#' @param type either \code{type = "raw"} to import the entire genome assembly 
+#' stats file or \code{type = "stats"} to import overall statistics including 
+#' all chromosomes, mitochondria and plastids.
 #' @author Hajk-Georg Drost
-#' @details This function takes a string specifying the path to the Genome Assembly Stats file
-#' of interest (e.g. the path returned by \code{\link{getAssemblyStats}}) and imports it.
-#' @seealso \code{\link{getAssemblyStats}}, \code{\link{read_genome}}, \code{\link{read_proteome}}, \code{\link{read_cds}}, \code{\link{read_gff}}
+#' @details This function takes a string specifying the path to the Genome 
+#' Assembly Stats file of interest (e.g. the path returned by 
+#' \code{\link{getAssemblyStats}}) and imports it.
+#' @seealso \code{\link{getAssemblyStats}}, \code{\link{read_genome}}, 
+#' \code{\link{read_proteome}}, \code{\link{read_cds}}, \code{\link{read_gff}}
 #' @export
+
 read_assemblystats <- function(file, type = "raw") {
     
+    
     if (!is.element(type, c("raw", "stats")))
-        stop("Please choose a type that is supported byt this function, e.g. type = 'raw' or type = 'stats'.", call. = FALSE)
+        stop(
+            "Please choose a type that is supported byt this function,
+            e.g. type = 'raw' or type = 'stats'.",
+            call. = FALSE
+        )
     
     unit_name <- molecule_name <- statistic <- NULL
     
@@ -38,46 +50,91 @@ read_assemblystats <- function(file, type = "raw") {
     
     if (type == "stats") {
         # select all features (including chromosomes, mitochondria, and plastids)
-        assemblystats_file.all.features <- dplyr::filter(assemblystats_file, unit_name == "all", molecule_name == "all")
+        assemblystats_file.all.features <-
+            dplyr::filter(assemblystats_file,
+                          unit_name == "all",
+                          molecule_name == "all")
         
-        total.length <- dplyr::filter(assemblystats_file.all.features, statistic == "total-length")
-        spanned.gaps <- dplyr::filter(assemblystats_file.all.features, statistic == "spanned-gaps")
-        unspanned.gaps <- dplyr::filter(assemblystats_file.all.features, statistic == "unspanned-gaps")
-        region.count <- dplyr::filter(assemblystats_file.all.features, statistic == "region-count")
-        scaffold.count <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-count")
-        scaffold.N50 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-N50")
-        scaffold.L50 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-L50")
-        scaffold.N75 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-N75")
-        scaffold.N90 <- dplyr::filter(assemblystats_file.all.features, statistic == "scaffold-N90")
-        contig.count <- dplyr::filter(assemblystats_file.all.features, statistic == "contig-count")
[…]
