+++ R/read_cds.R
@@ -1,121 +1,141 @@
 #' @title Import CDS as Biostrings or data.table object
-#' @description This function reads an organism specific CDS stored in a defined file format.
-#' @param file a character string specifying the path to the file storing the CDS.
-#' @param format a character string specifying the file format used to store the genome, e.g. \code{format = "fasta"} (default) or \code{format = "gbk"}.
-#' @param obj.type a character string specifying the object stype in which the genomic sequence shall be represented. 
-#' Either as \code{obj.type = "Biostrings"} (default) or as \code{obj.type = "data.table"}.
-#' @param delete_corrupt a logical value specifying whether potential CDS sequences that cannot be divided by 3 shall be
+#' @description This function reads an organism specific CDS stored in a 
+#' defined file format.
+#' @param file a character string specifying the path to the file storing 
+#' the CDS.
+#' @param format a character string specifying the file format used to store 
+#' the genome, e.g. \code{format = "fasta"} (default) or \code{format = "gbk"}.
+#' @param obj.type a character string specifying the object stype in which the 
+#' genomic sequence shall be represented. 
+#' Either as \code{obj.type = "Biostrings"} (default) or as 
+#' \code{obj.type = "data.table"}.
+#' @param delete_corrupt a logical value specifying whether potential CDS 
+#' sequences that cannot be divided by 3 shall be
 #' be excluded from the the dataset. Default is \code{delete_corrupt = FALSE}.
-#' @param ... additional arguments that are used by the seqinr::read.fasta() function.
+#' @param ... additional arguments that are used by
+#' \code{\link[seqinr]{read.fasta()}}.
 #' @author Hajk-Georg Drost
-#' @details The \code{read.cds} function takes a string specifying the path to the cds file
-#' of interest as first argument.
+#' @details The \code{read.cds} function takes a string specifying the path 
+#' to the cds file of interest as first argument.
 #'
-#' It is possible to read in different proteome file standards such as \emph{fasta} or \emph{genebank}.
+#' It is possible to read in different proteome file standards such as 
+#' \emph{fasta} or \emph{genebank}.
 #'
-#' CDS stored in fasta files can be downloaded from http://www.ensembl.org/info/data/ftp/index.html.
-#'
-#' @examples \dontrun{
-#' # reading a cds file stored in fasta format
-#' Ath.cds <- read_cds(system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
-#'                     format = "fasta")
-#' }
-#'
-#' @return A data.table storing the gene id in the first column and the corresponding
-#' sequence as string in the second column.
-#' @seealso \code{\link{getCDS}}, \code{\link{read_genome}}, \code{\link{read_proteome}}, \code{\link{read_gff}}, \code{\link{read_rna}}
+#' CDS stored in fasta files can be downloaded from 
+#' http://www.ensembl.org/info/data/ftp/index.html.
+#' @return A data.table storing the gene id in the first column and the 
+#' corresponding sequence as string in the second column.
+#' @seealso \code{\link{getCDS}}, \code{\link{read_genome}}, 
+#' \code{\link{read_proteome}}, \code{\link{read_gff}}, \code{\link{read_rna}}
 #' @import Biostrings
 #' @export
-read_cds <- function(file, format = "fasta", obj.type = "Biostrings", delete_corrupt = FALSE, ...){
-    
-    if (!is.element(format, c("fasta", "gbk")))
-        stop("Please choose a file format that is supported by this function.", call. = FALSE)
-    
-    if (!is.element(obj.type, c("Biostrings", "data.table")))
-        stop(
-            "Please specify a valid object type: obj.type = 'Biostrings' (default) or obj.type = 'data.table'.", call. = FALSE
-        )
-    
-    geneids <- seqs <- NULL
-    
[…]
