### 2017-04-08 3e6fdc8
change sapply -> lapply
