### 2017-04-08 e3114f3
1:ncol() -> seq_len()
