### 2017-04-08 c6461b2
1:ncol() -> seq_len()
