+++ R/getFilters.R
@@ -102,7 +102,7 @@ getFilters <- function(mart, dataset){
                              )), 1, function(x)
                                  unlist(strsplit(x, "\t")))))
         
-        colnames(rawDF) <- paste0("V", 1:ncol(rawDF))
+        colnames(rawDF) <- paste0("V", seq_len(ncol(rawDF)))
         
         filterBioMart <-
             as.data.frame(rawDF[, c("V1", "V2")],
