### 2017-04-08 9d4dda6
change sapply -> lapply
