### 2017-04-08 db182ac
change sapply -> lapply
