### 2017-04-08 e20543f
change sapply -> lapply
