### 2017-04-08 01d344e
change sapply -> lapply
