### 2017-04-08 8734d2d
change sapply -> lapply
