### 2017-04-08 41388ef
change sapply -> lapply
