### 2017-04-08 d48064c
change sapply -> lapply
