### 2017-04-08 864ddba
1:ncol() -> seq_len()
