### 2017-04-08 297557b
change sapply -> lapply
