### 2017-04-08 ef65a8f
change sapply -> lapply
