### 2017-04-08 d1a18dc
change sapply -> lapply
