+++ tests/testthat/test-download.database.R
@@ -1,7 +1,8 @@
 context("Test: download.database()")
 
 
-test_that("The download.database() throws error when wrong input database is specified..",{
+test_that("The download.database() throws error when wrong input
+          database is specified..",{
     
     skip_on_cran()
 
