+++ tests/testthat/test-getAttributes.R
@@ -10,10 +10,13 @@ test_that("The getAttributes() interface works properly..",{
         
         skip_on_cran()
         
-        marts <- getMarts()$mart
-        ensemb_mart <- which(stringr::str_match(marts, "ENSEMBL_MART_ENSEMBL") != "NA")
-        
-        attrib <- getAttributes(mart = as.character(marts[ensemb_mart]), dataset = "hsapiens_gene_ensembl")
-        
-            expect_equal(attrib[1 , 1], "ensembl_gene_id")
+    marts <- getMarts()$mart
+    ensemb_mart <-
+        which(stringr::str_match(marts, "ENSEMBL_MART_ENSEMBL") != "NA")
+    
+    attrib <-
+        getAttributes(mart = as.character(marts[ensemb_mart]), 
+                      dataset = "hsapiens_gene_ensembl")
+    
+    expect_equal(attrib[1 , 1], "ensembl_gene_id")
 })
