+++ tests/testthat/test-getDatasets.R
@@ -10,14 +10,21 @@ test_that("The getDatasets() interface works properly..",{
         
         skip_on_cran()
         
-        marts <- getMarts()[ , "mart"]
-        ex_mart <- which(stringr::str_match(marts, "ENSEMBL_MART_ENSEMBL") != "NA")
-        
-        attrib <- getDatasets(mart = as.character(marts$mart[ex_mart]))
-        
-        expect_true(equal_lists(as.list(attrib[1 , ]), list(dataset = "oanatinus_gene_ensembl",
-                                                            description = "Platypus genes (OANA5)",
-                                                            version = "OANA5")))
+    marts <- getMarts()[, "mart"]
+    ex_mart <-
+        which(stringr::str_match(marts, "ENSEMBL_MART_ENSEMBL") != "NA")
+    
+    attrib <-
+        getDatasets(mart = as.character(marts$mart[ex_mart]))
+    
+    expect_true(equal_lists(
+        as.list(attrib[1 , ]),
+        list(
+            dataset = "oanatinus_gene_ensembl",
+            description = "Platypus genes (OANA5)",
+            version = "OANA5"
+        )
+    ))
 })
 
 
