+++ tests/testthat/test-getKingdoms().R
@@ -3,8 +3,21 @@ context("Test: getKingdoms()")
 test_that("The getKingdoms() interface works properly..",{
     
     skip_on_cran()
-    skip_on_travis() 
-    testthat::equals(getKingdoms(), c("archaea", "bacteria", "fungi", "invertebrate", "plant", "protozoa", "vertebrate_mammalian", "vertebrate_other", "viral"))
-
+    skip_on_travis()
+    testthat::equals(
+        getKingdoms(),
+        c(
+            "archaea",
+            "bacteria",
+            "fungi",
+            "invertebrate",
+            "plant",
+            "protozoa",
+            "vertebrate_mammalian",
+            "vertebrate_other",
+            "viral"
+        )
+    )
+    
 })
 
