+++ tests/testthat/test-is.genome.available.R
@@ -1,50 +1,70 @@
 context("Test: is.genome.available()")
 
-test_that("The is.genome.available() interface works properly..",{
-        
-        skip_on_cran()
-        
-        # test whether interface to 'refseq' works properly
-        g <- is.genome.available(db = "refseq", organism = "Arabidopsis thaliana", details = TRUE)
-        expect_identical(g$organism_name,"Arabidopsis thaliana")
-        
-        # test with a second run using locally stored information
-        is.genome.available(organism = "Homo sapiens", details = TRUE)
-        
-        # test with a second run using locally stored information without details
-        is.genome.available(organism = "Homo sapiens")
-        
-        # test whether interface to 'genbank' works properly
-        is.genome.available(db = "genbank", organism = "Homo sapiens", details = TRUE)
-        
-        # test whether it works properly a second time based on the internal files
-        is.genome.available(db = "genbank", organism = "Arabidopsis thaliana", details = TRUE)
-        
-        # test whether it works properly a second time based on the internal files without details
-        is.genome.available(db = "genbank", organism = "Arabidopsis thaliana")
-        
-        # test whether interface to 'ensembl' works properly
-        is.genome.available(db = "ensembl", organism = "Homo sapiens", details = TRUE)
-        
-        # test whether interface to 'ensembl' works properly without details
-        is.genome.available(db = "ensembl", organism = "Homo sapiens")
-        
-        # test whether interface to 'ensemblgenomes' works properly
-        is.genome.available(db = "ensemblgenomes", organism = "Arabidopsis thaliana", details = TRUE)
-        
-        # test whether interface to 'ensemblgenomes' works properly without details
-        is.genome.available(db = "ensemblgenomes", organism = "Arabidopsis thaliana")
-        
+test_that("The is.genome.available() interface works properly..", {
+    skip_on_cran()
+    
+    # test whether interface to 'refseq' works properly
+    g <-
+        is.genome.available(db = "refseq",
+                            organism = "Arabidopsis thaliana",
+                            details = TRUE)
+    expect_identical(g$organism_name, "Arabidopsis thaliana")
+    
+    # test with a second run using locally stored information
+    is.genome.available(organism = "Homo sapiens", details = TRUE)
+    
+    # test with a second run using locally stored information without details
+    is.genome.available(organism = "Homo sapiens")
+    
+    # test whether interface to 'genbank' works properly
+    is.genome.available(db = "genbank",
+                        organism = "Homo sapiens",
+                        details = TRUE)
+    
+    # test whether it works properly a second time based on the internal files
+    is.genome.available(db = "genbank",
+                        organism = "Arabidopsis thaliana",
+                        details = TRUE)
+    
+    # test whether it works properly a second time based on the internal 
+    # files without details
+    is.genome.available(db = "genbank", organism = "Arabidopsis thaliana")
+    
+    # test whether interface to 'ensembl' works properly
+    is.genome.available(db = "ensembl",
+                        organism = "Homo sapiens",
+                        details = TRUE)
+    
+    # test whether interface to 'ensembl' works properly without details
+    is.genome.available(db = "ensembl", organism = "Homo sapiens")
+    
+    # test whether interface to 'ensemblgenomes' works properly
+    is.genome.available(db = "ensemblgenomes",
+                        organism = "Arabidopsis thaliana",
+                        details = TRUE)
+    
+    # test whether interface to 'ensemblgenomes' works properly without details
+    is.genome.available(db = "ensemblgenomes", 
+                        organism = "Arabidopsis thaliana")
+    
 })
 
 
[…]
