### 2017-04-10 60c1f7e
add package manual
