### 2017-04-10 aaff4d8
update NEWS
