### 2017-04-10 6c52649
now returns file path to make it easier for pipeline (%>%) construction
