### 2017-04-10 745a7c0
now returns file path to make it easier for pipeline (%>%) construction
