### 2017-04-10 c265787
update error messages
