### 2017-04-10 4da4f8c
update NEWS
