### 2017-04-10 b386439
add error handling
