+++ R/getAttributes.R
@@ -23,12 +23,19 @@
 #' \code{\link{getFilters}}, \code{\link{organismBM}}, 
 #' \code{\link{organismFilters}}, \code{\link{organismAttributes}}
 #' @export
+
 getAttributes <- function(mart, dataset){
     
     if ((!is.character(mart)) || (!is.character(dataset)))
         stop("Please use a character string as mart or dataset.",
              call. = FALSE)
+   
+    if (!is.element(mart, getMarts()$mart))
+        stop("Please select a valid mart with getMarts().", call. = FALSE)
     
+    if (!is.element(dataset, getDatasets(mart = mart)$dataset))
+        stop("Please select a valid dataset with getDatasets().", call. = FALSE)
+     
     if (stringr::str_detect(mart, "ENSEMBL"))
         # connect to BioMart API
         url <-
