### 2017-04-10 110decb
return paths of downloaded files
