### 2017-04-10 ddbbf8a
return paths of downloaded files
