+++ R/getENSEMBLInfo.R
@@ -2,6 +2,9 @@
 #' @description Retrieve species and genome information from 
 #' http://rest.ensembl.org/info/species?content-type=application/json/.
 #' @author Hajk-Georg Drost
+#' @examples 
+#' info.file <- getENSEMBLInfo()
+#' info.file
 #' @export 
 getENSEMBLInfo <- function() {
     ENSEMBLInfoTable <- get.ensembl.info(update = TRUE)
