+++ tests/testthat/test-getENSEMBLGENOMESInfo.R
@@ -1,0 +1,7 @@
+context("Test: getENSEMBLGENOMESInfo()")
+
+test_that("The getENSEMBLGENOMESInfo() interface works properly..",{
+    
+    # ENSEMBLGENOMES Info file retrieval
+    getENSEMBLGENOMESInfo()
+})
