+++ tests/testthat/test-getENSEMBLInfo.R
@@ -1,0 +1,7 @@
+context("Test: getENSEMBLInfo()")
+
+test_that("The getENSEMBLInfo() interface works properly..",{
+    
+    # ENSEMBLGENOMES Info file retrieval
+    getENSEMBLInfo()
+})
