### 2017-04-11 75add8c
add examples
