+++ R/getGENOMEREPORT.R
@@ -2,6 +2,9 @@
 #' @description Retrieves NCBI GENOME_REPORTS file from 
 #' ftp://ftp.ncbi.nlm.nih.gov/genomes/GENOME_REPORTS/overview.txt.
 #' @author Hajk-Georg Drost
+#' @examples 
+#' report <- getGENOMEREPORT()
+#' report
 #' @export
 
 getGENOMEREPORT <- function() {
