+++ man/getAssemblyStats.Rd
@@ -79,8 +79,8 @@ Scerevisiae.stats.import <- getAssemblyStats( db = "refseq",
 
 }
 \seealso{
-\code{\link{getProteome}}, \code{\link{getCDS}},
-\code{\link{getGFF}}, \code{\link{meta.retrieval}}, 
+\code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}},
+\code{\link{getGFF}}, \code{\link{getRNA}}, \code{\link{meta.retrieval}}, 
 \code{\link{read_assemblystats}}
 }
 \author{
