+++ tests/testthat/test-getGENOMEREPORT.R
@@ -1,0 +1,7 @@
+context("Test: getGENOMEREPORT()")
+
+test_that("The getGENOMEREPORT() interface works properly..",{
+    
+    # NCBI GENOMEREPORT Info file retrieval
+    getGENOMEREPORT()
+})
