+++ tests/testthat/test-getGFF.R
@@ -1,19 +1,33 @@
 context("Test: getGFF()")
 
-test_that("The getGFF() interface works properly for RefSeq..",{
+test_that("The getGFF() interface works properly for NCBI RefSeq..",{
     
     skip_on_cran()
 
     # test proper download from refseq
     getGFF( db       = "refseq", 
-                   organism = "Homo sapiens", 
+                   organism = "Saccharomyces cerevisiae", 
                    path = file.path("_ncbi_downloads","annotation"))
 })    
 
-test_that("The getGFF() interface works properly for Genbank..",{
+test_that("The getGFF() interface works properly for NCBI Genbank..",{
     # test proper download from genbank
     getGFF( db       = "genbank", 
-                   organism = "Homo sapiens", 
+                   organism = "Saccharomyces cerevisiae", 
                    path = file.path("_ncbi_downloads","annotation"))
 })    
 
+
+test_that("The getGFF() interface works properly for Ensembl",{
+    # test proper download from Ensembl
+    getGFF( db       = "ensembl", 
+            organism = "Saccharomyces cerevisiae", 
+            path = file.path("_ncbi_downloads","annotation"))
+})
+
+test_that("The getGFF() interface works properly for EnsemblGenomes",{
+    # test proper download from EnsemblGenomes
+    getGFF( db       = "ensemblgenomes", 
+            organism = "Saccharomyces cerevisiae", 
+            path = file.path("_ncbi_downloads","annotation"))
+})
