+++ R/getMetaGenomeAnnotations.R
@@ -26,10 +26,10 @@ getMetaGenomeAnnotations <-
              path = file.path("_ncbi_downloads", "metagenome", "annotations")) {
         if (!is.element(name, listMetaGenomes(details = FALSE)))
             stop(
-                "Unfortunately, the metagenome '",
+                paste0("Unfortunately, the metagenome '",
                 name,
-                "' is not available. Please consult the listMetaGenomes() 
-                function for available metagenomes."
+                "' is not available. Please consult the listMetaGenomes() ",
+                "function for available metagenomes.")
             )
         
         if (!file.exists(path)) {
