+++ man/getProteome.Rd
@@ -9,8 +9,13 @@ getProteome(db = "refseq", organism, path = file.path("_ncbi_downloads",
 }
 \arguments{
 \item{db}{a character string specifying the database from which the genome 
-shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, 
-\code{db = "ensembl"} or \code{db = "ensemblgenomes"}.}
+shall be retrieved:
+\itemize{
+\item \code{db = "refseq"}
+\item \code{db = "genbank"}
+\item \code{db = "ensembl"}
+\item \code{db = "ensemblgenomes"}
+}}
 
 \item{organism}{a character string specifying the scientific name of the 
 organism of interest, e.g. \code{organism = "Homo sapiens"}.}
