### 2017-04-11 f1f3748
update unit test
