### 2017-04-11 bc2a565
update unit test
