+++ tests/testthat/test-getMetaGenomeSummary.R
@@ -1,0 +1,5 @@
+context("Test: getMetaGenomeSummary()")
+
+test_that("The getMetaGenomeSummary() interface works properly..",{
+    getMetaGenomeSummary()
+})
