### 2017-04-11 #93 @HajkD
Dear @sckott,

I addressed all comments and issues and incorporated all changes into the new version of the `biomartr` package.

I hope that I could address all points sufficiently so that
you can consider `biomartr` to become a part of the `rOpenSci`   package collection.

Please find my detailed response below.

Thank you so much for all your support.

Kind regards,

Hajk

P.S. I just want to make you aware that the Ensembl Services are [currently
not available](https://ensembl.github.io/ensembl-status/index.html). Thus, until the Ensembl servers is up again some of the `biomartr`
functions will not run properly.

>For running egs, I suggest MAYBE running egs wrapped in dontrun on Travis by adding --run-dontrun to your check args (since taking a longish time on travis I assume is not a big deal) - I'd have one or two examples for each function not wrapped in dontrun when you know they will run quickly. Note: I said MAYBE since check already takes a very long time, but still would be nice to make sure egs always work

__Response__: I tried to solve the running time issue by choosing example test cases that
use very small files. It should be better now. Please also see my response to the test coverage and `R CMD check`.

>output of goodpractice::gp() was pointed out and there's some things in there which I think you should address:
sapply usage: replace with vapply/lapply
it's a bit of work, but I highly suggest tidying code and documentation to 80 line width so you don't get text running off the page on the CRAN manual (e.g., https://cran.rstudio.com/web/packages/biomartr/biomartr.pdf), and it's easier to see in a narrow width editor setup
replace usage of 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) with eg., seq_along/seq_len

__Response__: I agree and now replaced all `sapply()` commands by `lapply()`, reformatted all code and documentation to 80 line width, and replace usage of `1:length(...)`, `1:nrow(...)`, `1:ncol(...)`, `1:NROW(...)` and `1:NCOL(...)` with `seq_len()`.

>Add a package level manual file - so that users can do ?biomartr to get to the man pages easily

__Response__: A package level manual file is now included and as suggested can be accessed via `?biomartr`.

>looks like test coverage is about 47% - would be good to get that % up - not necessary now, but in the future.

__Response__: I now spent a significant amount of time on increasing the test coverage and will continue to improve it in the future.

>R cmd check takes a long time, just for ease of quick turn around time, it'd be wise to make that time much less - I assume most of the time is in tests, but haven't looked into how to speed that up

__Response__: Most tests cover the functionality of data retrieval functions, so to reduce `R cmd check` time I now choose small file downloads for unit testing.

>also related to tests, but not about speed: looks like many of your test files aren't doing much testing using testthat - that is, you run examples, but then don't test their output (e.g., https://github.com/HajkD/biomartr/blob/master/tests/testthat/test-is.genome.available.R#L12-L36) I imagine the majority of time in tests right now is doing data requests - adding test expectations shouldn't add a lot more time

__Response__: I absolutely agree and see where your confusion comes from. I now commented and specified the unit tests more clearly, so that no confusion exists on what part of the function is actually tested.
