### 2017-04-11 2714ffa
adding md5 check
