### 2017-04-11 422cafe
helper function to read md5 files
