### 2017-04-13 d6ae17c
add global variable
