### 2017-04-13 fe80fbe
add global variable
