### 2017-04-13 83a91c0
add global variable
