### 2017-04-13 84ef3f3
add global variable
