### 2017-04-13 98d4c0f
add global variable
