+++ tests/testthat/test-getGFF.R
@@ -3,7 +3,7 @@ context("Test: getGFF()")
 test_that("The getGFF() interface works properly for NCBI RefSeq..",{
     
     skip_on_cran()
-
+        skip_on_travis()
     # test proper download from refseq
     getGFF( db       = "refseq", 
                    organism = "Saccharomyces cerevisiae", 
@@ -11,6 +11,8 @@ test_that("The getGFF() interface works properly for NCBI RefSeq..",{
 })    
 
 test_that("The getGFF() interface works properly for NCBI Genbank..",{
+        skip_on_travis()
+        skip_on_cran()
     # test proper download from genbank
     getGFF( db       = "genbank", 
                    organism = "Saccharomyces cerevisiae", 
@@ -19,6 +21,8 @@ test_that("The getGFF() interface works properly for NCBI Genbank..",{
 
 
 test_that("The getGFF() interface works properly for Ensembl",{
+        skip_on_cran()
+        skip_on_travis()
     # test proper download from Ensembl
     getGFF( db       = "ensembl", 
             organism = "Saccharomyces cerevisiae", 
@@ -26,6 +30,8 @@ test_that("The getGFF() interface works properly for Ensembl",{
 })
 
 test_that("The getGFF() interface works properly for EnsemblGenomes",{
+        skip_on_cran()
+        skip_on_travis()
     # test proper download from EnsemblGenomes
     getGFF( db       = "ensemblgenomes", 
             organism = "Saccharomyces cerevisiae", 
