+++ tests/testthat/test-getKingdomAssemblySummary.R
@@ -3,6 +3,7 @@ context("Test: getKingdomAssemblySummary()")
 test_that("The getKingdomAssemblySummary() interface works properly
           for NCBI RefSeq..",{
     skip_on_cran()
+    skip_on_travis()             
     # KingdomAssembly Summary Info file retrieval from RefSeq
     getKingdomAssemblySummary("refseq")
 })
@@ -11,6 +12,7 @@ test_that("The getKingdomAssemblySummary() interface works properly
 test_that("The getKingdomAssemblySummary() interface works properly
           for NCBI Genbank..",{
               skip_on_cran()
+              skip_on_travis()          
               # KingdomAssembly Summary Info file retrieval from Genbank
               getKingdomAssemblySummary("genbank")
 })
