### 2017-04-15 bca03fd
update unit test
