### 2017-04-17 #93 @sckott
@grimbough @naupaka Are you happy with changes made?  Anything else?

Looking over your changes now @HajkD
