### 2017-04-17 #93 @naupaka
I probably won't have a chance to take a look till later this week. I'll make another pass through then to make sure all works as expected on my machine - seems like all the major points were addressed.
