### 2017-04-17 #93 @sckott
@HajkD Looks like check passes cleanly, examples aren't too bad - two files that can be faster are definitely `organismAttributes` and to lesser extent `organismFilters`

tests still take a very long time - eventually try to get test time down

there's a failure in one of the tests

```r
2: getENSEMBL.Annotation(organism, type = "dna", id.type = "toplevel", path)
     3: is.element(stringr::str_to_lower(new.organism), ensembl.available.organisms$name)
     4: match(el, set, 0L)

     trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/metagenomes/assembly_summary.txt'
     ftp data connection made, file length 333931 bytes
     ==================================================
     downloaded 326 KB

     OK: 50 SKIPPED: 21 FAILED: 1
     1. Error: The getGFF() interface works properly for Ensembl (@test-getGFF.R#27)
```
