+++ FAQs.md
@@ -1,5 +1,6 @@
 # FAQs
 
+- [How is the `biomartr` package different from the `BiomaRt` package that I have been using so far?](https://github.com/HajkD/biomartr/issues/11)
 - [How can I retrieve GO terms for human genes using ensembl gene IDs?](https://github.com/HajkD/biomartr/issues/5)
 - [How can I download coding sequences or protein sequences for all sequenced species?](https://www.biostars.org/p/9202/#235962)
 - [Where can I Download the human reference genome in fasta format?](https://www.biostars.org/p/1796/#236039)
