### 2017-04-19 da707e3
speed up example runs
