### 2017-04-19 17df1f8
speed up example runs
