+++ R/organismFilters.R
@@ -32,8 +32,6 @@
 #' "_biomart/listFilters_organism.txt", allowing subsequent queries to perform 
 #' much faster.
 #' @examples 
-#' # return available filters for "Homo sapiens"
-#' head(organismFilters("Homo sapiens"), 20)
 #' # search for filter topic "id" 
 #' head(organismFilters("Homo sapiens", topic = "id"), 20)
 #' @references
