### 2017-04-19 #93 @HajkD
Dear @sckott

Thank you very much for having another look at `biomartr`.

I now reduced the examples for `organismAttributes` and `organismFilters` to make the files faster. Now there is only one example each. However, it still takes quite long. So either I wrap the examples with `dontrun` or we leave it as it is. What do you think?

Would it be possible that you re-run the tests on your machine because I couldn't reproduce the test-error of `getGFF` that you posted? Sometimes, functions need to be re-run if the internet connection is not stable enough. I am aware that the unit test execution time is partly responsible for that and I will do my best to work on that.

In the future, I will try to reduce the execution time of the unit tests. However, it will always remain to be a trade-off between unit test execution time and test coverage due to the large file size retrieval functionality of the package. I am always open and grateful for suggestions :) .

I am very much looking forward to receiving further feedback and comments from @naupaka and @grimbough.

Kind regards,
Hajk
