### 2017-04-19 #93 @sckott
yep, i'll rerun tests and report back
