### 2017-04-19 #93 @sckott
same error again. DO you get that test error on travis. Have you tried Appveyor?  Or r-hub ?
