### 2017-04-19 #93 @HajkD
Hm, that's strange. Yes, Travis passes. I just committed today -> https://travis-ci.org/HajkD/biomartr/builds/223596728 . I haven't tried rhub or Appveyor yet, but I can run it there as well if you like.
