### 2017-04-22 #93 @HajkD
Dear @sckott

I now also ran an r-hub check and everything seems to pass: https://builder.r-hub.io/status/biomartr_0.5.0.tar.gz-e5116cc3ef554c2db801cbbaa3be1e0f .
