### 2017-04-25 #93 @naupaka
@HajkD thanks for making many of the recommended changes. Most all of the things I mentioned the first time around are now fixed. I'm still having issues with getting things to work smoothly with `devtools::check()` though. I think perhaps it is related to the timeout problem you mentioned - but it seems to happen in the same place pretty consistently. Might there be a way to fail gracefully if it's not your fault the server is timing out and then still have the tests pass?

Unless maybe this is something insidious that snuck in because of my recent update to R 3.4.0?

My system info is below, and I was using commit ca36885e2 on master for these tests.

```
> sessionInfo()
R version 3.4.0 (2017-04-21)
Platform: x86_64-apple-darwin16.5.0 (64-bit)
Running under: macOS Sierra 10.12.4

Matrix products: default
BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libLAPACK.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

other attached packages:
[1] biomartr_0.5.0

 [1] Rcpp_0.12.10         IRanges_2.9.19       XML_3.98-1.6         Biostrings_2.43.8    digest_0.6.12
 [6] bitops_1.0-6         DBI_0.6-1            stats4_3.4.0         RSQLite_1.1-2        zlibbioc_1.21.0
[11] curl_2.5             XVector_0.15.2       S4Vectors_0.13.17    tools_3.4.0          Biobase_2.35.1
[16] biomaRt_2.31.10      RCurl_1.95-4.8       parallel_3.4.0       compiler_3.4.0       BiocGenerics_0.21.3
[21] AnnotationDbi_1.37.4 memoise_1.1.0
```

Output from `devtools::check()`:
---
```r
devtools::check(document = FALSE)

Setting env vars ---------------------------------------------------------------
CFLAGS  : -Wall -pedantic
CXXFLAGS: -Wall -pedantic
Building biomartr --------------------------------------------------------------
'/usr/local/Cellar/r/3.4.0/R.framework/Resources/bin/R' --no-site-file  \
  --no-environ --no-save --no-restore --quiet CMD build  \
  '/Users/naupaka/git/ROpenSci/biomartr' --no-resave-data --no-manual

* checking for file ‘/Users/naupaka/git/ROpenSci/biomartr/DESCRIPTION’ ... OK
* preparing ‘biomartr’:
* creating vignettes ... OK
* checking for LF line-endings in source and make files
* checking for empty or unneeded directories
* building ‘biomartr_0.5.0.tar.gz’

Setting env vars ---------------------------------------------------------------
_R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
_R_CHECK_CRAN_INCOMING_           : FALSE
_R_CHECK_FORCE_SUGGESTS_          : FALSE
Checking biomartr --------------------------------------------------------------
'/usr/local/Cellar/r/3.4.0/R.framework/Resources/bin/R' --no-site-file  \
  --no-environ --no-save --no-restore --quiet CMD check  \
  '/var/folders/1y/8s_b2fx520sgj5f3j9fjwg880000gn/T//Rtmpvxn83G/biomartr_0.5.0.tar.gz'  \
  --as-cran --timings --no-manual

* using log directory ‘/Users/naupaka/git/ROpenSci/biomartr.Rcheck’
* using R version 3.4.0 (2017-04-21)
* using platform: x86_64-apple-darwin16.5.0 (64-bit)
* using session charset: UTF-8
* using options ‘--no-manual --as-cran’
* checking for file ‘biomartr/DESCRIPTION’ ... OK
* this is package ‘biomartr’ version ‘0.5.0’
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘biomartr’ can be installed ... OK
* checking installed package size ... OK
* checking package directory ... OK
* checking ‘build’ directory ... OK
* checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘biomartr_0.5.0.tar.gz’
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking R files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... OK
* checking whether the package can be loaded with stated dependencies ... OK
* checking whether the package can be unloaded cleanly ... OK
* checking whether the namespace can be loaded with stated dependencies ... OK
* checking whether the namespace can be unloaded cleanly ... OK
* checking loading without being on the library search path ... OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... OK
* checking Rd files ... OK
* checking Rd metadata ... OK
* checking Rd line widths ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
* checking installed files from ‘inst/doc’ ... OK
* checking files in ‘vignettes’ ... OK
* checking examples ... ERROR
Running examples in ‘biomartr-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: organismAttributes
> ### Title: Retrieve Ensembl Biomart attributes for a query organism
> ### Aliases: organismAttributes
>
> ### ** Examples
>
> # search for attribute topic id
> head(organismAttributes("Homo sapiens", topic = "id"), 20)
Warning: No attributes were available for mart = ENSEMBL_MART_SEQUENCE and dataset = hsapiens_genomic_sequence.
Warning: No attributes were available for mart = ENSEMBL_MART_SEQUENCE and dataset = hsapiens_genomic_sequence.
Error in curl::curl_fetch_memory(url, handle = handle) :
  Timeout was reached
Calls: head ... request_fetch -> request_fetch.write_memory -> <Anonymous> -> .Call
Execution halted
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
  Running ‘testthat.R’ [2m/19m]
 ERROR
Last 13 lines of output:
  trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/metagenomes/assembly_summary.txt'
  Content type 'unknown' length 333931 bytes (326 KB)
  ==================================================
  testthat results ================================================================
  OK: 48 SKIPPED: 0 FAILED: 7
  1. Error: The biomart() interface works properly.. (@test-biomart.R#10)
  2. Error: The getAttributes() interface works properly.. (@test-getAttributes.R#13)
  3. Error: The getCDS() interface to Ensembl works properly.. (@test-getCDS.R#26)
  4. Error: The getCDS() interface to EnsemblGenomes works properly.. (@test-getCDS.R#37)
  5. Error: The getCDS() error messages work properly.. (@test-getCDS.R#47)
  6. Error: The getDatasets() interface works properly.. (@test-getDatasets.R#13)
  7. Error: The getGO() interface works properly.. (@test-getGO.R#12)

  Error: testthat unit tests failed
  Execution halted
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes in ‘inst/doc’ ... OK
* checking re-building of vignette outputs ... WARNING
Error in re-building vignettes:
  ...
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/fungi/assembly_summary.txt'
Content type 'unknown' length 79196 bytes (77 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/invertebrate/assembly_summary.txt'
Content type 'unknown' length 43167 bytes (42 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/plant/assembly_summary.txt'
Content type 'unknown' length 26815 bytes (26 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/protozoa/assembly_summary.txt'
Content type 'unknown' length 24012 bytes (23 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_mammalian/assembly_summary.txt'
Content type 'unknown' length 32978 bytes (32 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_other/assembly_summary.txt'
Content type 'unknown' length 37450 bytes (36 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/viral/assembly_summary.txt'
Content type 'unknown' length 1748600 bytes (1.7 MB)
==================================================
Quitting from lines 371-379 (Functional_Annotation.Rmd)
Error: processing vignette 'Functional_Annotation.Rmd' failed with diagnostics:
Request to BioMart web service failed. Verify if you are still connected to the internet.  Alternatively the BioMart web service is temporarily down.
Execution halted

* DONE
Status: 2 ERRORs, 1 WARNING, 1 NOTE

See
  ‘/Users/naupaka/git/ROpenSci/biomartr.Rcheck/00check.log’
for details.

checking examples ... ERROR
Running examples in ‘biomartr-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: organismAttributes
> ### Title: Retrieve Ensembl Biomart attributes for a query organism
> ### Aliases: organismAttributes
>
> ### ** Examples
>
> # search for attribute topic id
> head(organismAttributes("Homo sapiens", topic = "id"), 20)
Warning: No attributes were available for mart = ENSEMBL_MART_SEQUENCE and dataset = hsapiens_genomic_sequence.
Warning: No attributes were available for mart = ENSEMBL_MART_SEQUENCE and dataset = hsapiens_genomic_sequence.
Error in curl::curl_fetch_memory(url, handle = handle) :
  Timeout was reached
Calls: head ... request_fetch -> request_fetch.write_memory -> <Anonymous> -> .Call
Execution halted

checking tests ... ERROR
  Running ‘testthat.R’ [2m/19m]
Last 13 lines of output:
  trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/metagenomes/assembly_summary.txt'
  Content type 'unknown' length 333931 bytes (326 KB)
  ==================================================
  testthat results ================================================================
  OK: 48 SKIPPED: 0 FAILED: 7
  1. Error: The biomart() interface works properly.. (@test-biomart.R#10)
  2. Error: The getAttributes() interface works properly.. (@test-getAttributes.R#13)
  3. Error: The getCDS() interface to Ensembl works properly.. (@test-getCDS.R#26)
  4. Error: The getCDS() interface to EnsemblGenomes works properly.. (@test-getCDS.R#37)
  5. Error: The getCDS() error messages work properly.. (@test-getCDS.R#47)
  6. Error: The getDatasets() interface works properly.. (@test-getDatasets.R#13)
  7. Error: The getGO() interface works properly.. (@test-getGO.R#12)

  Error: testthat unit tests failed
  Execution halted

checking re-building of vignette outputs ... WARNING
Error in re-building vignettes:
  ...
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/fungi/assembly_summary.txt'
Content type 'unknown' length 79196 bytes (77 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/invertebrate/assembly_summary.txt'
Content type 'unknown' length 43167 bytes (42 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/plant/assembly_summary.txt'
... 8 lines ...
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_other/assembly_summary.txt'
Content type 'unknown' length 37450 bytes (36 KB)
==================================================
trying URL 'ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/viral/assembly_summary.txt'
Content type 'unknown' length 1748600 bytes (1.7 MB)
==================================================
Quitting from lines 371-379 (Functional_Annotation.Rmd)
Error: processing vignette 'Functional_Annotation.Rmd' failed with diagnostics:
Request to BioMart web service failed. Verify if you are still connected to the internet.  Alternatively the BioMart web service is temporarily down.
Execution halted

checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘biomartr_0.5.0.tar.gz’
R CMD check results
2 errors | 1 warning  | 1 note

R CMD check succeeded
```
