+++ NAMESPACE
@@ -13,6 +13,7 @@ export(getFilters)
 export(getGENOMEREPORT)
 export(getGFF)
 export(getGO)
+export(getGTF)
 export(getGenome)
 export(getGroups)
 export(getKingdomAssemblySummary)
