### 2017-05-10 3d9d7cc
update NEWS
