+++ R/getCDS.R
@@ -47,7 +47,9 @@ getCDS <-
                 'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.",
                 call. = FALSE
             )
-        
+            
+        message("Starting retrieval of ", organism," ...")
+            
         if (is.element(db, c("refseq", "genbank"))) {
             # get Kingdom Assembly Summary file
             AssemblyFilesAllKingdoms <-
