### 2017-05-10 2e03635
add new internal helper function for GTF retrieval
