### 2017-05-10 c95c509
add new internal helper function for GTF retrieval
