+++ R/getENSEMBLGENOMES.gtf.R
@@ -1,0 +1,403 @@
+#' @title Annotation file retrieval from ENSEMBLGENOMES
+#' @description Helper function for retrieving GFF files
+#' of a particular organism of interest from the ENSEMBLGENOMES ftp server
+#' @param organism scientific name of the query organism
+#' @param type type of biological sequence
+#' @param id.type ENSEMBLGENOMES id type
+#' @param path where shall file be saved?
+#' @author Hajk-Georg Drost
+#' @noRd
+
+getENSEMBLGENOMES.gtf <-
+        function(organism,
+                 type = "dna",
+                 id.type = "toplevel",
+                 path) {
+                
+                if (!is.element(type, c("dna", "cds", "pep")))
+                        stop("Please a 'type' argument supported by this function: 
+                             'dna', 'cds', 'pep'.")
+                
+                # test if REST API is responding
+                is.ensemblgenomes.alive()
+                
+                name <- NULL
+                
+                new.organism <- stringr::str_replace_all(organism, " ", "_")
+                
+                if (file.exists(file.path(tempdir(), "ensemblgenomes_summary.txt"))) {
+                        suppressWarnings(
+                                ensembl.available.organisms <-
+                                        readr::read_tsv(
+                                                file.path(tempdir(), "ensemblgenomes_summary.txt"),
+                                                col_names = c(
+                                                        "division",
+                                                        "taxon_id",
+                                                        "name",
+                                                        "release",
+                                                        "display_name",
+                                                        "accession",
+                                                        "common_name",
+                                                        "assembly"
+                                                ),
+                                                col_types = readr::cols(
+                                                        division = readr::col_character(),
+                                                        taxon_id = readr::col_integer(),
+                                                        name = readr::col_character(),
+                                                        release = readr::col_integer(),
+                                                        display_name = readr::col_character(),
+                                                        accession = readr::col_character(),
+                                                        common_name = readr::col_character(),
+                                                        assembly = readr::col_character()
+                                                ),
+                                                comment = "#"
+                                        )
+                        )
+                }
+                
+                if (!file.exists(file.path(tempdir(), "ensemblgenomes_summary.txt"))) {
+                        # check if organism is available on ENSEMBL
+                        tryCatch({
+                                ensembl.available.organisms <-
+                                        jsonlite::fromJSON(
+                                                "http://rest.ensemblgenomes.org/info/species?content-type=application/json"
+                                        )
+                        }, error = function(e)
+                                stop(
+                                        "The API 'http://rest.ensemblgenomes.org' does not seem to 
+                                        work properly. Are you connected to the internet? 
+                                        Is the homepage 'http://rest.ensemblgenomes.org' 
+                                        currently available?",
+                                        call. = FALSE
+                                ))
+                        
+                        aliases <- groups <- NULL
+                        
+                        # transform list object returned by 'fromJSON' to tibble
+                        ensembl.available.organisms <-
+                                tibble::as_tibble(dplyr::select(
+                                        ensembl.available.organisms$species,
+                                        -aliases,
+                                        -groups
+                                ))
+                        
+                        readr::write_tsv(
+                                ensembl.available.organisms,
+                                file.path(tempdir(), "ensemblgenomes_summary.txt")
+                        )
+                }
+                
+                if (!any(
+                        stringr::str_detect(
+                                stringr::str_to_lower(new.organism),
+                                ensembl.available.organisms$name
+                        )
+                )) {
+                        warning(
+                                "Unfortunately organism '",
+                                organism,
+                                "' is not available at ENSEMBL. Please check whether or not the 
+                                organism name is typed correctly. Thus, download of this species
+                                has been omitted."
+                        )
+                        return(FALSE)
+                }
[…]
