+++ R/getENSEMBLGENOMESInfo.R
@@ -2,9 +2,11 @@
 #' @description Retrieve species and genome information from 
 #' http://rest.ensemblgenomes.org/info/species?content-type=application/json/.
 #' @author Hajk-Georg Drost
-#' @examples 
+#' @examples
+#' \dontrun{ 
 #' info.file <- getENSEMBLGENOMESInfo()
 #' info.file
+#' }
 #' @export 
 getENSEMBLGENOMESInfo <- function() {
     ENSEMBLGENOMESInfoTable <- get.ensemblgenome.info(update = TRUE)
