### 2017-05-10 2d027a5
add new function for GTF retrieval from ensembl and ensemblgenomes
