+++ R/getProteome.R
@@ -61,7 +61,9 @@ getProteome <-
                 'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.",
                 call. = FALSE
             )
-        
+            
+        message("Starting retrieval of ", organism," ...")
+            
         if (is.element(db, c("refseq", "genbank"))) {
             # get Kingdom Assembly Summary file
             AssemblyFilesAllKingdoms <-
