+++ R/getRNA.R
@@ -26,8 +26,8 @@
 #' Ath_RNA <- read_rna(file_path, format = "fasta")
 #' }
 #' @seealso \code{\link{getGenome}}, \code{\link{getProteome}}, 
-#' \code{\link{getRNA}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, 
-#' \code{\link{read_cds}}
+#' \code{\link{getGTF}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, 
+#' \code{\link{read_cds}}, \code{\link{getCDS}}
 #' @export
 
 getRNA <-
@@ -41,7 +41,9 @@ getRNA <-
                 'refseq', 'genbank', 'ensembl' or 'ensemblgenomes'.",
                 call. = FALSE
             )
-        
+            
+        message("Starting retrieval of ", organism, " ...")
+            
         if (is.element(db, c("refseq", "genbank"))) {
             # get Kingdom Assembly Summary file
             AssemblyFilesAllKingdoms <-
