+++ man/getENSEMBLGENOMESInfo.Rd
@@ -11,8 +11,10 @@ Retrieve species and genome information from
 http://rest.ensemblgenomes.org/info/species?content-type=application/json/.
 }
 \examples{
+\dontrun{ 
 info.file <- getENSEMBLGENOMESInfo()
 info.file
+}
 }
 \author{
 Hajk-Georg Drost
