+++ man/getRNA.Rd
@@ -42,8 +42,8 @@ Ath_RNA <- read_rna(file_path, format = "fasta")
 }
 \seealso{
 \code{\link{getGenome}}, \code{\link{getProteome}}, 
-\code{\link{getRNA}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, 
-\code{\link{read_cds}}
+\code{\link{getGTF}}, \code{\link{getGFF}}, \code{\link{meta.retrieval}}, 
+\code{\link{read_cds}}, \code{\link{getCDS}}
 }
 \author{
 Hajk-Georg Drost
