### 2017-06-07 #93 @HajkD
Hi @sckott,

Please accept my apologies for my late response. I was caught up in another project that required my immediate attention.

Nevertheless, I worked on the remaining `biomartr` issues and released the new version 0.5.1 on CRAN.

While digging deeper into the issue I noticed that the NCBI and ENSEMBL databases don't provide very stable APIs. Servers are frequently maintained and sometimes down for several days. Hence, it is impossible to resolve this issue from my side. However, I now implemented warnings that will occur when connections are not stable or whenever the server is down.

I am very happy for any suggestions on how I can improve the `ftp` connection via R to the NCBI and ENSEMBL servers.

In future versions, I also plan to implement `https` and `rsync` access to the NCBI and ENSEMBL servers, but these options will exclude Windows users. So the initial `ftp` issues will always exist for Windows users.

I now run `rhub::check(platform = rhub::platforms()$name,  valgrind = TRUE)` whenever I check `biomartr` and don't receive errors.

When running `devtools::check()` on my machine, I receive the following output:

```r
* using log directory ‘/private/var/folders/3x/6bbw6ds1039gpwny1m0hn8r80000gp/T/RtmpEQJf3b/biomartr.Rcheck’
* using R version 3.4.0 (2017-04-21)
* using platform: x86_64-apple-darwin15.6.0 (64-bit)
* using session charset: UTF-8
* using options ‘--no-manual --as-cran’
* checking for file ‘biomartr/DESCRIPTION’ ... OK
* this is package ‘biomartr’ version ‘0.5.9000’
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘biomartr’ can be installed ... OK
* checking installed package size ... OK
* checking package directory ... OK
* checking ‘build’ directory ... OK
* checking top-level files ... OK
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking R files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... OK
* checking whether the package can be loaded with stated dependencies ... OK
* checking whether the package can be unloaded cleanly ... OK
* checking whether the namespace can be loaded with stated dependencies ... OK
* checking whether the namespace can be unloaded cleanly ... OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... OK
* checking Rd files ... OK
* checking Rd metadata ... OK
* checking Rd line widths ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
* checking installed files from ‘inst/doc’ ... OK
* checking files in ‘vignettes’ ... OK
* checking examples ... OK
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
  Running ‘testthat.R’ [95s/460s]
 OK
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes in ‘inst/doc’ ... OK
* checking re-building of vignette outputs ... OK
* DONE

Status: OK

R CMD check results
0 errors | 0 warnings | 0 notes
```

I would be truly grateful if you could test again on your system if the check runs smoothly now.

I hope that you can now consider `biomartr` to be integrated into the `rOpenSci` package collection and I am happy to continue working on it and improving it once the onboarding process was successful.

I would like to thank all of you for your time a detailed code review. It helped me a lot to significantly improve the `biomartr` functionality.

Kind regards,
Hajk
