### 2017-06-07 #93 @HajkD
Dear @sckott

I am very grateful for your approval and I very much look forward to collaborating with you and the rOpenSci community.

I already transferred the package and would be very interested in a blog post for the more general audience (https://ropensci.org/blog/).

Thank you so much for coordinating everything. I truly appreciate it.

Kind regards,
Hajk
