### 2017-06-07 #93 @sckott
Great!

- Make sure to update any links from `HajkD` to `ropensci`
- for the blog post, here's instructions https://github.com/ropensci/roweb/blob/master/.github/CONTRIBUTING.md#blog-posts - we'll get back to you soon with a proposed date when it can go up on our site - some egs of what previous folks have done with blog post about onboarded pkgs: https://ropensci.github.io/biweekly/update-2017-06-05/#onboarding-series
