### 2017-06-20 #93 @HajkD
Many thanks, @sckott. I am quite busy in the past weeks, but I will try to write a post by today. So far, your directions are very clear, but I will happily take your offer to help me in case I get stuck :)
