### 2017-06-20 #93 @stefaniebutland
Hello @HajkD. Just wanted to say how happy I am that you will contribute a post about `biomartr`. In my (past) research life I used BioMart, especially as an interface to Ensembl annotations, and preached about its utility to my biologist colleagues. Great to have this addition to rOpenSci!
