### 2017-06-20 #93 @HajkD
Hello @stefaniebutland ,

Thank you so much for your great feedback :) I truly hope that this package will become a small puzzle stone to help to tackle the reproducibility crisis in biomedical research and also to speed up large-scale genomics studies. If it's useful then I am happy about any collaboration to extend the `biomartr` functionality :)
