### 2017-06-22 #93 @HajkD
Hi @sckott ,

I am very sorry that I still didn't manage to find some quiet time to write a proper blog post. I am quite caught up in a paper revision and collaborations right now.
Would it be fine if I postpone my submission by another week? If not, please feel free to onboard `biomartr` without a blog entry and I am happy to contribute to a blog post later on (if that's also fine with you).

Many thanks for your patience and support :)

Kind regards,
Hajk
