### 2017-06-22 #93 @sckott
@HajkD another week is fine, no rush at all!

actually, @stefaniebutland when could we fit his post in the blog schedule?
