### 2017-06-22 #93 @stefaniebutland
@HajkD `biomartr` can def be onboarded without a post and your paper revision is important!

Could you give me an idea of a month and week that would work for you? I'll work with that to assign you a date to provide a draft and post date. Cheers.
