### 2017-06-25 #93 @HajkD
Dear @sckott and @stefaniebutland ,

Thank you so much for your prompt reply and for all your support!

@stefaniebutland: Would the week between 03.07. - 09.07. work for you? I could deliver a blog post on any day during that week.

I truly appreciate all your efforts and for helping me to schedule the post :)

Kind regards,
Hajk
