### 2017-06-26 #93 @stefaniebutland
@HajkD A draft in the week of 03.07. - 09.07 would work well. Tentative post date would be Tues July 18. This would give us time to give you feedback on your draft post.
Cheers
