### 2017-06-27 #93 @HajkD
Dear @stefaniebutland

Many thanks. This sounds perfect. I will post my draft between 03.07. - 09.07 and very much look forward to receiving your feedback.

Kind regards,
Hajk
